% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_modules.R
\name{accuracyModuleUI}
\alias{accuracyModuleUI}
\title{Accuracy gauge - UI side}
\usage{
accuracyModuleUI(id)
}
\arguments{
\item{id}{character, module id, unique name per plot}
}
\value{
ui object
}
\description{
Accuracy gauge - UI side
}
\author{
mvarewyck
}
