% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxAgeGenderLowerJaw.R
\name{ageGenderLowerJawServer}
\alias{ageGenderLowerJawServer}
\title{Shiny module for creating the plot \code{\link{boxAgeGenderLowerJaw}} - server side}
\usage{
ageGenderLowerJawServer(id, data, types, timeRange)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{types}{defines the species types that can be selected}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{boxAgeGenderLowerJaw}} - server side
}
\author{
mvarewyck
}
