% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barBiotoop.R
\name{barBiotoop}
\alias{barBiotoop}
\title{Create interactive plot for biotoop description - irrespective of species}
\usage{
barBiotoop(data, jaar = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame, background data for the WBE, as read from \code{loadHabitats}}

\item{jaar}{numeric, year of interest}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given year a summary of the WBE parameters
defining the percentage of land, water, forest etc}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'Naam': }{characteristic}
\item{'Value': }{value}
}
}
}
}
\description{
Create interactive plot for biotoop description - irrespective of species
}
\author{
mvarewyck
}
