% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barDraagkracht.R
\name{barDraagkracht}
\alias{barDraagkracht}
\title{Function to generate barplots for  Maatschappelijk draagvlak (F12_1)}
\usage{
barDraagkracht(
  data,
  groupVariable = NULL,
  xVar = "percentage",
  yVar = NULL,
  width = 1000,
  height = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{groupVariable}{character vector containing the name of the column in plotData
for which each unique value a separate subplot is created}

\item{xVar}{character, column in \code{plotData} for x-axis;
default value is 'percentage'}

\item{yVar}{character, column in \code{plotData} for y-axis}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with plotly object and data.frame
}
\description{
Function to generate barplots for  Maatschappelijk draagvlak (F12_1)
}
\author{
wverlinden
}
