% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barDraagkracht.R
\name{barDraagkrachtServer}
\alias{barDraagkrachtServer}
\title{Shiny module for creating the plot \code{\link{barDraagkracht}} - server side}
\usage{
barDraagkrachtServer(id, data, groupVariable = NULL, xVar = "percentage", yVar)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{groupVariable}{character vector containing the name of the column in plotData
for which each unique value a separate subplot is created}

\item{xVar}{character, column in \code{plotData} for x-axis;
default value is 'percentage'}

\item{yVar}{character, column in \code{plotData} for y-axis}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{barDraagkracht}} - server side
}
\author{
mvarewyck
}
