% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barInschatting.R
\name{barInschatting}
\alias{barInschatting}
\title{Function to generate barplot for Inschatting van het risico op verkeersongelukken (F07_3)}
\usage{
barInschatting(data)
}
\arguments{
\item{data}{dataframe with columns including percentage, Vraag and Antwoord}
}
\value{
list with plotly object and data.frame
}
\description{
Function to generate barplot for Inschatting van het risico op verkeersongelukken (F07_3)
}
\author{
wverlinden
}
