% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxAgeGenderLowerJaw.R
\name{boxAgeGenderLowerJaw}
\alias{boxAgeGenderLowerJaw}
\title{Create interactive plot for comparing length of lower jaw between age/gender category}
\usage{
boxAgeGenderLowerJaw(data, wildNaam = "", jaartallen = NULL, regio = "",
  width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object with length lower jaw per 
age/gender category for specified \code{jaartallen}}
\item{'data': }{raw data used for the plot, as data.frame with:
\itemize{
\item{'ageGender': }{age/gender category}
\item{'onderkaaklengte': }{length of the lower jaw in mm}
}
}
}
}
\description{
Figure p.17 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
\author{
Laure Cougnaud
}
