% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxAgeWeight.R
\name{boxAgeWeight}
\alias{boxAgeWeight}
\title{Create interactive plot for distribution of weight in function of age}
\usage{
boxAgeWeight(data, wildNaam = "", jaartallen = NULL, regio = "",
  width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object, for a given species the distribution of weight
in function of age is plotted in box plots
}
\description{
Figure p. 17 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
