% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxAgeWeight.R
\name{boxAgeWeightServer}
\alias{boxAgeWeightServer}
\title{Shiny module for creating the plot \code{\link{boxAgeWeight}} - server side}
\usage{
boxAgeWeightServer(id, data, type, timeRange)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{type}{animal type, used to filter \code{data},
based on 'Leeftijdscategorie_onderkaak' column}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{boxAgeWeight}} - server side
}
\author{
mvarewyck
}
