% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxAgeWeight.R
\name{boxAgeWeightUI}
\alias{boxAgeWeightUI}
\title{Shiny module for creating the plot \code{\link{boxAgeWeight}} - UI side}
\usage{
boxAgeWeightUI(id, uiText)
}
\description{
Shiny module for creating the plot \code{\link{boxAgeWeight}} - UI side
}
\author{
mvarewyck
}
