% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxRealisedShot.R
\name{boxRealisedShot}
\alias{boxRealisedShot}
\title{Create interactive boxplot for percentage verwezenlijkt per jaar}
\usage{
boxRealisedShot(
  data,
  regio,
  type = NULL,
  jaartallen = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame, with \code{loadToekenningen()} for specific WBE}

\item{regio, }{empty function argument needed for generalization in \code{\link{plotModuleServer}}}

\item{type}{animal type, used to filter \code{data} ('type_comp' column)
default is \code{c("Smalree", "Geit")}}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for the specified type and jaartallen}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'jaar': }{year at which the animal was shot}
\item{'verwezenlijkt': }{count for actually shot animals}
\item{'toegekend': }{count for assigned animals}
\item{'percent': }{verwezenlijkt / toegekend * 100}
}}
}
}
\description{
Create interactive boxplot for percentage verwezenlijkt per jaar
}
\author{
mvarewyck
}
