% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxRealisedShot.R
\name{boxRealisedShotUI}
\alias{boxRealisedShotUI}
\title{Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side}
\usage{
boxRealisedShotUI(id, showAccuracy = FALSE, regionLevels = NULL, uiText)
}
\arguments{
\item{showAccuracy}{boolean, whether to show gauge for accuracy}

\item{regionLevels}{numeric vector, region level choices}
}
\description{
Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side
}
\author{
mvarewyck
}
