% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeCheek.R
\name{countAgeCheek}
\alias{countAgeCheek}
\title{Create interactive plot for comparing age based on cheek or reported by hunter}
\usage{
countAgeCheek(
  data,
  jaartallen = NULL,
  currentYear = as.numeric(format(Sys.Date(), "\%Y")),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{currentYear}{numeric, current year to calculate accuracy for}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given species the percentage per age category
based on cheek or hunter report in a stacked bar chart}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'jager': }{category based on the meldingsformulier}
\item{'kaak': }{category based on the lower jaw inspection}
\item{'freq': }{counts of animals}
\item{'percent': }{percentage in the entire category based on
the lower jaw inspection}
}
}
}
}
\description{
Figure p. 9 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
\author{
mvarewyck
}
