% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeCheek.R
\name{countAgeCheekServer}
\alias{countAgeCheekServer}
\title{Shiny module for creating the plot \code{\link{countAgeCheek}} - server side}
\usage{
countAgeCheekServer(id, data, timeRange)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countAgeCheek}} - server side
}
\author{
mvarewyck
}
