% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeCheek.R
\name{countAgeCheekUI}
\alias{countAgeCheekUI}
\title{Shiny module for creating the plot \code{\link{countAgeCheek}} - UI side}
\usage{
countAgeCheekUI(id, showAccuracy = FALSE, uiText)
}
\arguments{
\item{showAccuracy}{boolean, whether to show gauge for accuracy}
}
\description{
Shiny module for creating the plot \code{\link{countAgeCheek}} - UI side
}
\author{
mvarewyck
}
