% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeGender.R
\name{countAgeGenderServer}
\alias{countAgeGenderServer}
\title{Shiny module for creating the plot \code{\link{countAgeGender}} - server side}
\usage{
countAgeGenderServer(id, data, timeRange)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot, see \code{\link{countAgeGender}}}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countAgeGender}} - server side
}
\author{
mvarewyck
}
