% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeGroup.R
\name{countAgeGroup}
\alias{countAgeGroup}
\title{Create interactive plot for comparing age and group variable}
\usage{
countAgeGroup(data, groupVariable, jaartallen = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{groupVariable}{character, variable in \code{data}}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given species the percentage per age category
and per group, based on meldingsformulier data in a stacked bar chart}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'leeftijd': }{age category}
\item{'group': }{group indicator}
\item{'freq': }{counts of animals}
\item{'percent': }{percentage of counts of animals}
}
}
}
}
\description{
Create interactive plot for comparing age and group variable
}
\author{
mvarewyck
}
