% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeGroup.R
\name{countAgeGroupUI}
\alias{countAgeGroupUI}
\title{Shiny module for creating the plot \code{\link{countAgeGroup}} - UI side}
\usage{
countAgeGroupUI(id, uiText, doHide = TRUE)
}
\arguments{
\item{doHide}{boolean, whether to initially hide the plot; default TRUE}
}
\description{
Shiny module for creating the plot \code{\link{countAgeGroup}} - UI side
}
\author{
mvarewyck
}
