% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEmbryos.R
\name{countEmbryos}
\alias{countEmbryos}
\title{Create interactive plot for number of embryos versus year}
\usage{
countEmbryos(
  data,
  type = c("Smalree", "Reegeit"),
  jaartallen = NULL,
  regio = "",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{type}{animal type, used to filter \code{data} ('ageGender' column)
default is \code{c("Smalree", "Geit")}}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for the specified specie and years}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'afschotjaar': }{year at which the animal was shot}
\item{'variable': }{age/gender category}
\item{'Freq': }{counts of females}
\item{'percent': }{percentage of females with given number of embryos per year}
}}
}
}
\description{
Adapted version from Figure p. 30 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
\author{
mvarewyck
}
