% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEmbryos.R
\name{countEmbryosServer}
\alias{countEmbryosServer}
\title{Shiny module for creating the plot \code{\link{countEmbryos}} - UI side}
\usage{
countEmbryosServer(id, data, timeRange, types, uiText, wildsoort)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{types}{defines the species types that can be selected}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{wildsoort}{character, species to be displayed. 
Needed to format title and description in \code{uiText}}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countEmbryos}} - UI side
}
\author{
mvarewyck
}
