% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEmbryos.R
\name{countEmbryosUI}
\alias{countEmbryosUI}
\title{Shiny module for creating the plot \code{\link{countEmbryos}} - UI side}
\usage{
countEmbryosUI(id, regionLevels, uiText)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{regionLevels}{character, choices for region}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{countEmbryos}} - UI side
}
\author{
mvarewyck
}
