% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countHuntingMethod.R
\name{countHuntingMethod}
\alias{countHuntingMethod}
\title{Create interactive plot for percentage per hunting method}
\usage{
countHuntingMethod(
  data,
  regio,
  locaties,
  jaartallen,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{regio}{character, regional level of interest should be one of 
\code{c("provinces", "flanders", "faunabeheerzones")}}

\item{locaties}{character vector, selected set of \code{regio} to be retained}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given specie the observed number 
per year and per province is plotted in a stacked bar chart}
\item{'data': }{data.frame, summary data for download}
}
}
\description{
Create interactive plot for percentage per hunting method
}
