% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countHuntingMethod.R
\name{countHuntingMethodUI}
\alias{countHuntingMethodUI}
\title{Shiny module for creating the plot \code{\link{countHuntingMethod}} - UI side}
\usage{
countHuntingMethodUI(id, regionLevels = NULL)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{regionLevels}{numeric vector, region level choices}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{countHuntingMethod}} - UI side
}
\author{
mvarewyck
}
