% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoten_indicatorIntroductionPathway.R
\name{countIntroductionPathway}
\alias{countIntroductionPathway}
\title{Plot number of new introductions per pathway at level 2}
\usage{
countIntroductionPathway(data)
}
\arguments{
\item{data}{data.frame with raw data for plotting}
}
\value{
list with plot and data that was used for the plot (i.e. without missing values for the )
}
\description{
Split up by pathways at level 1
}
