% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearAge.R
\name{countYearAge}
\alias{countYearAge}
\title{Create interactive plot for counts per age category and year}
\usage{
countYearAge(data, wildNaam = "", jaartallen = NULL, regio = "",
  doodsoorzaak = "afschot", width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{doodsoorzaak}{character vector, defines the cause(s) of death that should
be considered in the plot}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object, for a given species the observed number 
per year and per age category is plotted in a stacked bar chart
}
\description{
Figure p. 11 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
