% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearAge.R
\name{countYearAgeUI}
\alias{countYearAgeUI}
\title{Shiny module for creating the plot \code{\link{countYearAge}} - UI side}
\usage{
countYearAgeUI(
  id,
  uiText,
  plotFunction = "countYearAgeUI",
  showRegion = TRUE,
  doHide = TRUE
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{showRegion}{boolean, whether to show the region filter; default is TRUE}

\item{doHide}{boolean, whether to initially hide the plot; default TRUE}

\item{title}{character, plot title in the app, overrides the automatic title}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{countYearAge}} - UI side
}
\author{
mvarewyck
}
