% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoten_indicatorYearByNativeRange.R
\name{countYearNativerange}
\alias{countYearNativerange}
\title{Create interactive plot for counts per native region and year of introduction}
\usage{
countYearNativerange(
  data,
  jaartallen = NULL,
  type = c("native_continent", "native_range"),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{type}{character, native_range level of interest should be one of 
\code{c("native_continent", "native_range")}}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given specie the observed number 
per year and per province is plotted in a stacked bar chart}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'year': }{year at which the animal was introduced}
\item{'nativeRange': }{native range of the introduced animal}
\item{'aantal': }{counts of animals}
}
}
}
}
\description{
Based on \code{\link{countYearProvince}} plot from grofwild
}
