% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearProvince.R
\name{countYearProvince}
\alias{countYearProvince}
\title{Create interactive plot for counts per province and year}
\usage{
countYearProvince(
  data,
  jaartallen = NULL,
  type = c("provinces", "flanders", "faunabeheerzones"),
  sourceIndicator = NULL,
  title = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{type}{character, regional level of interest should be one of 
\code{c("provinces", "flanders", "faunabeheerzones")}}

\item{sourceIndicator}{character, source used to filter \code{data} ('indieningType' column)
should be subset of \code{c("E-loket", "HVV", "Natuurpunt")}.}

\item{title}{character, title prefix; default is NULL}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given specie the observed number 
per year and per province is plotted in a stacked bar chart}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'afschotjaar': }{year at which the animals was shot}
\item{'locatie': }{location name, could be province, flanders or fbz name}
\item{'value': }{counts of animals}
}
}
}
}
\description{
Figure p. 4 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
