% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearProvince.R
\name{countYearProvinceUI}
\alias{countYearProvinceUI}
\title{Shiny module for creating the plot \code{\link{countYearProvince}} - UI side}
\usage{
countYearProvinceUI(id, uiText)
}
\description{
Shiny module for creating the plot \code{\link{countYearProvince}} - UI side
}
\author{
mvarewyck
}
