% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearRealizedShot.R
\name{countYearRealizedShot}
\alias{countYearRealizedShot}
\title{Create interactive plot for verwezenlijkt and toegekend afschot per jaar}
\usage{
countYearRealizedShot(
  data,
  assignedData,
  type = NULL,
  jaartallen = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{assignedData}{data.frame with summary data on the number of assigned
animals that can be shot}

\item{type}{animal type, used to filter \code{data} ('ageGender' column)
default is \code{c("Smalree", "Geit")}}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for the specified type and jaartallen}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'jaar': }{year at which the animal was shot}
\item{'verwezenlijkt': }{count for actually shot animals}
\item{'toegekend': }{count for assigned animals}
\item{'percent': }{verwezenlijkt / toegekend * 100}
}}
}
}
\description{
Create interactive plot for verwezenlijkt and toegekend afschot per jaar
}
\author{
mvarewyck
}
