% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearRegion.R
\name{countYearRegion}
\alias{countYearRegion}
\title{Create interactive plot for counts selected communes and years}
\usage{
countYearRegion(data, locaties = NULL, timeRange = NULL,
  width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{locaties}{character vector, regions that were selected to plot}

\item{timeRange}{numeric vector, time range selected for plot}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given species the observed number 
per year and per selected commune is plotted in a line plot}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'afschotjaar': }{year at which the animals was shot}
\item{'locatie': }{comune name}
\item{'freq': }{counts of animals}
}
}
}
}
\description{
Create interactive plot for counts selected communes and years
}
