% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearSchade.R
\name{countYearSchade}
\alias{countYearSchade}
\title{Create interactive barplot for wildschade data variable of interest ifo year}
\usage{
countYearSchade(
  data,
  jaartallen = NULL,
  type = NULL,
  summarizeBy = c("count", "percent"),
  fullNames = NULL,
  sourceIndicator = NULL,
  width = 800,
  height = 600
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{type}{character, variable name in \code{data} of interest}

\item{summarizeBy}{character, whether to summarize data in terms of counts or percentages}

\item{fullNames}{named character vector, values for the \code{variable} to be 
displayed instead of original data values}

\item{sourceIndicator}{character, source used to filter \code{data} ('indieningType' column)
should be subset of \code{c("E-loket", "HVV", "Natuurpunt")}.}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with
\itemize{
\item{'plot': }{plotly object}
\item{'data': }{data.fram used for plot}
}
}
\description{
Create interactive barplot for wildschade data variable of interest ifo year
}
\author{
mvarewyck
}
