% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearShotAnimals.R
\name{countYearShotAnimals}
\alias{countYearShotAnimals}
\title{Interactive barplot to show the distribution of the hunted animals in a certain interval}
\usage{
countYearShotAnimals(
  data,
  regio,
  jaartallen = NULL,
  width = NULL,
  height = NULL,
  interval = c("Per maand", "Per seizoen", "Per twee weken"),
  type = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{regio, }{empty function argument needed for generalization in \code{\link{plotModuleServer}}}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}

\item{interval}{character, data shown in intervals monthly, biweekly..}

\item{type}{character, used to filter the data}
}
\description{
Interactive barplot to show the distribution of the hunted animals in a certain interval
}
\author{
dbemelmans
}
