% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearShotAnimals.R
\name{countYearShotServer}
\alias{countYearShotServer}
\title{Shiny module for creating the plot \code{\link{countYearShotAnimals}} - server side}
\usage{
countYearShotServer(id, data, timeRange, types)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{types}{character vector}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countYearShotAnimals}} - server side
}
\author{
mvarewyck
}
