% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot2.R
\name{createPlot2}
\alias{createPlot2}
\title{Create interactive version of plot at page 7}
\usage{
createPlot2(width = NULL, height = NULL)
}
\arguments{
\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object
}
\description{
Create interactive version of plot at page 7
}
