% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlots.R
\name{createPlot3}
\alias{createPlot3}
\title{Create interactive version of plot at page 13}
\usage{
createPlot3(year = 2014, width = NULL, height = NULL)
}
\arguments{
\item{year}{integer, year that you want to compare to all remaining years}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object
}
\description{
Create interactive version of plot at page 13
}
