% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{createSchadeSummaryData}
\alias{createSchadeSummaryData}
\title{Summarized data for perceelplot}
\usage{
createSchadeSummaryData(schadeData, timeRange)
}
\arguments{
\item{schadeData}{spatialPointsDataFrame contains the points where there was
wildschade and descriptives in data.frame}

\item{timeRange}{numeric vector, year span of interest}
}
\value{
a filtered spatialPointsDataFrame
}
\description{
Create object of type schadedata, filtered for cases within \code{timeRange}
whilst retaining only a select number of columns relevant for the perceelPlot 
created by \code{\link{mapSchade}} and relevant for data download.
}
\author{
Eva Adriaensen
}
