% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{createSpaceData}
\alias{createSpaceData}
\title{Create summary data of geographical data for selected year, species and region level}
\usage{
createSpaceData(
  data,
  allSpatialData,
  biotoopData,
  year,
  species,
  regionLevel,
  unit = c("absolute", "relative", "absoluteCases", "relativeDekking"),
  sourceIndicator = NULL,
  dataDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{data}{data.frame, geographical data}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{biotoopData}{data.frame, background data for the WBE, as read from \code{loadHabitats}}

\item{year}{integer, year of interest}

\item{species}{character, species of interest}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes", "utm5" )}}

\item{unit}{character, whether absolute frequencies, relative frequencies (aantal/100ha),
absolute cases, or relative bos freq (aantal/100ha bos & natuur)  
should be reported,}

\item{sourceIndicator}{character, source used to filter \code{data} ('indieningType' column)
should be one of \code{c("E-loker", "HVV", "Natuurpunt")}.}

\item{dataDir}{character, path to data files}
}
\value{
a list with two items: data - a data.frame with the summary data; stats - a data.frame with the summary statistics
}
\description{
Create summary data of geographical data for selected year, species and region level
}
\author{
mvarewyck
}
