% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_create.R
\name{createTrafficData}
\alias{createTrafficData}
\title{Create shape data for dashboard wild zwijn - verkeer (traffic)}
\usage{
createTrafficData(
  jsonDir = "~/git/reporting-rshiny-grofwildjacht/data",
  dataDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{jsonDir}{character, path to json shape files}

\item{dataDir}{character, path to data files}
}
\value{
boolean, whether file is successfully saved
}
\description{
Create shape data for dashboard wild zwijn - verkeer (traffic)
}
\author{
mvarewyck
}
