% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{createTrendData}
\alias{createTrendData}
\title{Create data for plotting trend over years in \code{\link{trendYearFlanders}}}
\usage{
createTrendData(
  data,
  allSpatialData,
  timeRange,
  species,
  regionLevel,
  unit = c("absolute", "relative")
)
}
\arguments{
\item{data}{data.frame, geographical data}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{timeRange}{numeric vector, time range selected for plot}

\item{species}{character, species of interest}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes")}}

\item{unit}{character, whether absolute or relative frequencies (aantal/100ha) 
should be reported}
}
\value{
data.frame, summary of number of animals per species, region and year.
Ready for plotting with \code{\link{trendYearFlanders}}
}
\description{
Create data for plotting trend over years in \code{\link{trendYearFlanders}}
}
\author{
mvarewyck
}
