% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_create.R
\name{createWaarnemingenData}
\alias{createWaarnemingenData}
\title{Enrich waarnemingen data with CELLCODE and gemeentecode}
\usage{
createWaarnemingenData(
  dataFile = "~/git/reporting-rshiny-grofwildjacht/data/waarnemingen_2018.csv",
  saveDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{dataFile}{path to read current waarnemeningen data}

\item{saveDir}{path to write new waarnemingen data}
}
\value{
TRUE if creation succeeded
}
\description{
Enrich waarnemingen data with CELLCODE and gemeentecode
}
\author{
mvarewyck
}
