% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoten_indicatorIntroductionCumulative.R
\name{cumulativeIntroductionYear}
\alias{cumulativeIntroductionYear}
\title{Cumulative plot of yearly number of new introductions}
\usage{
cumulativeIntroductionYear(data, region = NULL)
}
\arguments{
\item{data}{}
}
\value{
list with plot and data
}
\description{
Cumulative plot of yearly number of new introductions
}
