% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{datatableModuleUI}
\alias{datatableModuleUI}
\title{Interactive table generated with datatable (ui-side)}
\usage{
datatableModuleUI(id)
}
\arguments{
\item{id}{character, module id, unique name per plot}
}
\value{
ui object
}
\description{
Interactive table generated with datatable (ui-side)
}
\author{
Eva Adriaensen
}
