% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{exoten_dutchNames}
\alias{exoten_dutchNames}
\title{List Dutch names to replace English names for exoten}
\usage{
exoten_dutchNames(x, type = c("regio"))
}
\arguments{
\item{x}{character, what to transform}

\item{type}{character, defines how to transform short to full names}
}
\value{
named character vector, names are the original values
}
\description{
Will not return NA, but rather the original English name in case
no match could be found.
}
\author{
eadriaensen
}
