% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{figureModuleServer}
\alias{figureModuleServer}
\title{User input for controlling specific plot (server-side)}
\usage{
figureModuleServer(input, output, session, data)
}
\arguments{
\item{input}{shiny input variable}

\item{output}{shiny output variable}

\item{session}{shiny session variable}

\item{data}{reactive data.frame, data for specific plot}
}
\value{
no return value; some output objects are created
}
\description{
User input for controlling specific plot (server-side)
}
