% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{figureModuleUI}
\alias{figureModuleUI}
\title{User input for controlling specific plot (ui-side)}
\usage{
figureModuleUI(id, title, showLegend = TRUE, showExtraVariables = TRUE,
  showTime = TRUE, showRegion = TRUE, showGlobe = TRUE)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{title}{character, title for the control panel}

\item{showLegend}{boolean, whether to show input field for the legend}

\item{showExtraVariables}{boolean, whether to show input field for 
extra variables (popup)}

\item{showTime}{boolean, whether to show input field for time range}

\item{showRegion}{boolean, whether to show input fields for region selection}

\item{showGlobe}{boolean, whether to show input field for map of the globe}
}
\value{
ui object (tagList)
}
\description{
User input for controlling specific plot (ui-side)
}
