% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{filterGrofwild}
\alias{filterGrofwild}
\title{Filter \code{plotData} based on leeftijd and geslacht bron}
\usage{
filterGrofwild(
  plotData,
  sourceIndicator_leeftijd = NULL,
  sourceIndicator_geslacht = NULL,
  sourceIndicator_onderkaak = NULL,
  sourceIndicator_embryos = NULL,
  returnStop = c("message", "data")
)
}
\arguments{
\item{plotData}{data.frame, to be filtered}

\item{sourceIndicator_leeftijd}{character, levels of \code{leeftijd_comp_bron} to select on}

\item{sourceIndicator_geslacht}{character, levels of \code{geslacht_comp_bron} to select on}

\item{sourceIndicator_onderkaak}{character, levels of \code{onderkaaklengte_comp_bron} to select on}

\item{sourceIndicator_embryos}{character, levels of \code{aantal_embryos_bron} to select on}

\item{returnStop}{character, should be one of \code{c("message", "data")}
what needs to be returned if the filtered data has no rows left}
}
\value{
data.frame, filtered version of \code{plotData}
}
\description{
Filter \code{plotData} based on leeftijd and geslacht bron
}
\author{
mvarewyck
}
