% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{filterSchade}
\alias{filterSchade}
\title{Filter \code{plotData} based on the \code{indieningType} if required}
\usage{
filterSchade(
  plotData,
  sourceIndicator = NULL,
  returnStop = c("message", "data")
)
}
\arguments{
\item{plotData}{data.frame, to be filtered}

\item{sourceIndicator}{character, source used to filter \code{data} ('indieningType' column)
should be subset of \code{c("E-loket", "HVV", "Natuurpunt")}.}

\item{returnStop}{character, should be one of \code{c("message", "data")}
what needs to be returned if the filtered data has no rows left}
}
\value{
data.frame, filtered version of \code{plotData}
}
\description{
Filter \code{plotData} based on the \code{indieningType} if required
}
\author{
mvarewyck
}
