% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{filterSpatial}
\alias{filterSpatial}
\title{Filter loaded \code{allSpatialData} for selected species, regionLevel and year}
\usage{
filterSpatial(
  allSpatialData,
  species,
  regionLevel = c("flanders", "provinces", "communes", "faunabeheerzones",
    "fbz_gemeentes", "utm5", "WBE", "WBE_buitengrenzen"),
  year,
  locaties = NULL
)
}
\arguments{
\item{allSpatialData}{list with SpatialPolygonsDataFrame as loaded by 
\code{load(file = file.path(system.file("extdata", package = "reportingGrofwild"), "spatialData.RData"))}}

\item{species}{character, animal species}

\item{regionLevel}{character, region level. Should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes", "utm5", "WBE_binnengrenzen")}}

\item{year}{integer, year of interest. Only relevant when \code{regionLevel}
is "WBE_binnengrenzen". For all other regionlevels spatial data is fixed over the years}

\item{locaties}{character vector, only relevant when \code{regionLevel} is 
"WBE_binnengrenzen"; it selects the relevant WBE only; default is NULL}
}
\value{
single SpatialPolygonsDataFrame
}
\description{
Filter loaded \code{allSpatialData} for selected species, regionLevel and year
}
\author{
mvarewyck
}
