% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{filterSpatialWbe}
\alias{filterSpatialWbe}
\title{Filter loaded \code{allSpatialData} for selected partijNummer}
\usage{
filterSpatialWbe(allSpatialData, partijNummer)
}
\arguments{
\item{allSpatialData}{list with SpatialPolygonsDataFrame as loaded by 
\code{load(file = file.path(system.file("extdata", package = "reportingGrofwild"), "spatialData.RData"))}}

\item{partijNummer}{numeric, partijnummer of the WBE to filter}
}
\value{
list with SpatialPolygonsDataFrame, each of them filtered on selected WBE
}
\description{
Filter loaded \code{allSpatialData} for selected partijNummer
}
\author{
mvarewyck
}
