% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{formatSchadeSummaryData}
\alias{formatSchadeSummaryData}
\title{Format summary schadedata data for download with nice column names and correct column order}
\usage{
formatSchadeSummaryData(summarySchadeData)
}
\arguments{
\item{summarySchadeData}{spatialPointsDataFrame as obtained from \code{\link{createSchadeSummaryData}}}
}
\value{
data.frame, dataframe with nicely formatted columnnames and specific column order
}
\description{
Format summary schadedata data for download with nice column names and correct column order
}
\author{
Eva Adriaensen
}
