% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fullNames}
\alias{fullNames}
\title{List user-friendly full names to replace R coding names}
\usage{
fullNames(x, type = "maanden", rev = FALSE)
}
\arguments{
\item{x}{character, what to transform}

\item{type}{character, defines how to transform short to full names}

\item{rev}{boolean, whether to switch names and values}
}
\value{
character vector
}
\description{
Will not return NA, but rather the R coding name in case
no match could be found.
}
\author{
mvarewyck
}
