% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPathFigure}
\alias{getPathFigure}
\title{get path of file contained in the 'figure' folder of the report}
\usage{
getPathFigure(figureName)
}
\arguments{
\item{figureName}{string name of the figure file
('graph[1-3].png')}
}
\value{
string with path of figure
}
\description{
get path of file contained in the 'figure' folder of the report
}
\author{
Laure Cougnaud
}
