% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{getProvince}
\alias{getProvince}
\title{Define province based on commune NIS code}
\usage{
getProvince(NISCODE, allSpatialData)
}
\arguments{
\item{NISCODE}{character vector, NIS codes from communes, for which to define province}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}
}
\value{
character vector, same length and order as \code{NISCODE}
with corresponding province for each communeS
}
\description{
Define province based on commune NIS code
}
\author{
mvarewyck
}
