% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getSeason}
\alias{getSeason}
\title{Define season from date}
\usage{
getSeason(dates)
}
\arguments{
\item{dates}{date vector, format \code{"d/m/Y"}}
}
\value{
character vector with respective season
}
\description{
Define season from date
}
\author{
mvarewyck
}
