% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoten_loadData.R
\name{loadExotenData}
\alias{loadExotenData}
\title{Read exoten data}
\usage{
loadExotenData(
  dataDir = system.file("extdata", package = "reportingGrofwild"),
  type = c("indicators")
)
}
\arguments{
\item{dataDir}{character vector, defines the path to the data file(s)}

\item{type}{data type, "indicators" for indicator data}
}
\value{
data.table, loaded indicator data; 
and attribute 'Date', the date that this data file was created
}
\description{
By default data for which \code{first_observed < 1950} is excluded.
}
