% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadHabitats}
\alias{loadHabitats}
\title{Load Habitats (Background) data}
\usage{
loadHabitats(
  dataDir = system.file("extdata", package = "reportingGrofwild"),
  spatialData,
  regionLevels = NULL
)
}
\arguments{
\item{dataDir}{character, path to data files}

\item{spatialData}{list with each element a SpatialPolygonsDataFrame as created 
by \code{\link{createShapeData}}}

\item{regionLevels}{character vector, for which regions load the habitat data;
if NULL loaded for all levels; default value is NULL}
}
\value{
named list with data.frame for each region level
}
\description{
Load Habitats (Background) data
}
\author{
mvarewyck
}
