% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadMetaEco}
\alias{loadMetaEco}
\title{Specify currently used levels in eco data}
\usage{
loadMetaEco(clean = TRUE)
}
\value{
list with meta data for eco data
}
\description{
Specify currently used levels in eco data
}
\author{
mvarewyck
}
