% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadMetaSchade}
\alias{loadMetaSchade}
\title{Specify currently used type schades}
\usage{
loadMetaSchade()
}
\value{
character vector
}
\description{
Specify currently used type schades
}
\author{
mvarewyck
}
