% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadRawData}
\alias{loadRawData}
\title{Read ecology, geography, wildschade, kbo_wbe or waarnemingen data}
\usage{
loadRawData(
  bucket = config::get("bucket", file = system.file("config.yml", package =
    "reportingGrofwild")),
  type = c("eco", "geo", "wildschade", "kbo_wbe")
)
}
\arguments{
\item{bucket}{character, name of the S3 bucket as specified in the config.yml file;
default value is "inbo-wbe-uat-data"}

\item{type}{data type, "eco" for ecology data and "geo" for geography data}
}
\value{
data.frame, loaded data
}
\description{
Data is preprocessed by createRawData() at INBO
}
\author{
mvarewyck
}
