% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadShapeData}
\alias{loadShapeData}
\title{Load all shape data from zipped files}
\usage{
loadShapeData(showProgress = FALSE)
}
\arguments{
\item{showProgress}{boolean, whether to show progress window in the shiny app;
Note, if used outside shiny app, this will cause an error; if FALSE progress
is printed in the console}
}
\value{
list with for each spatial level a SpatialPolygonsDataFrame object, 
with polygons and data as provided in the zipFile; spatial levels are 
flanders, provinces, communes and provincesVoeren (Voeren as separate province)
}
\description{
Load all shape data from zipped files
}
