% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadSpreadData}
\alias{loadSpreadData}
\title{Load data for \code{mapSpread}}
\usage{
loadSpreadData(
  spatialDir = system.file("extdata", package = "reporting-grofwild"),
  spatialLevel = c("pixels", "municipalities"),
  unit = c("model_EP", "model_OH", "risk_EP", "risk_OH")
)
}
\arguments{
\item{spatialDir}{path to directory with spatial files}

\item{spatialLevel}{character, for which spatial level to create the map;
should be one of \code{c("pixels", "municipalities")}}

\item{unit}{character, characteristic that defines the polygons and color coding
should be one of \code{c("model_EP", "model_OH", "risk_EP", "risk_OH")}}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Load data for \code{mapSpread}
}
\author{
mvarewyck
}
