% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadToekenningen}
\alias{loadToekenningen}
\title{Read toekenningen (Ree) data}
\usage{
loadToekenningen()
}
\value{
data.frame with columns:
\itemize{
\item{'Provincie': }{character, province}
\item{'Jaar': }{integer, year}
\item{'Labeltype': }{character, type of Ree, one of \code{c("Geiten", "Bokken", "Kitsen")}}
\item{'Aantal': }{integer, frequencie per categorie}
}
}
\description{
Read toekenningen (Ree) data
}
