% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadWbeHabitats}
\alias{loadWbeHabitats}
\title{Load WBE Habitats (Background) data}
\usage{
loadWbeHabitats(
  dataDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{dataDir}{character vector, defines the path to the data files}
}
\value{
data.frame
}
\description{
Load WBE Habitats (Background) data
}
\author{
mvarewyck
}
