% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{mapFlandersServer}
\alias{mapFlandersServer}
\title{Shiny module for creating the plot \code{\link{mapFlanders}} - server side}
\usage{
mapFlandersServer(
  id,
  defaultYear,
  species,
  currentWbe = NULL,
  hideGlobeDefault = TRUE,
  isSchade = FALSE,
  geoData,
  biotoopData = NULL,
  allSpatialData
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{defaultYear}{numeric, default year}

\item{species}{character, species for which to show the graphs}

\item{currentWbe}{numeric, KBO number; default value is NULL}

\item{hideGlobeDefault}{boolean, whether the globe is shown by default 
when the map is first created; default value is TRUE}

\item{geoData}{SpatialPolygonsDataFrame with geographical data}

\item{biotoopData}{data.frame, with background biotoop data for selected region level;
default value is NULL}

\item{allSpatialData}{list with SpatialPolygonsDataFrame}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{mapFlanders}} - server side
}
\author{
mvarewyck
}
