% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{mapFlandersUI}
\alias{mapFlandersUI}
\title{Shiny module for creating the plot \code{\link{mapFlanders}} - UI side}
\usage{
mapFlandersUI(
  id,
  showRegion = TRUE,
  showSource = FALSE,
  showCombine = TRUE,
  type = c("grofwild", "wildschade", "wbe", "dash"),
  unitChoices = c(Aantal = "absolute", `Aantal/100ha` = "relative"),
  plotDetails = c("flanders", "region")
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{showRegion}{boolean, whether to show choices for regionLevel and selected region(s)}

\item{showSource}{boolean, whether to show choices for sources}

\item{showCombine}{boolean, whether to show option to combine selected regions}

\item{type}{character, defines the layout depending on which page it is shown;
should be one of \code{c("grofwild", "wildschade", "wbe", "dash")}}

\item{unitChoices}{named character vector, choices for unit option;
default is \code{c("Aantal" = "absolute", "Aantal/100ha" = "relative")}}

\item{plotDetails}{character vector, detail plots to be shown below the map;
should be subset of \code{c("flanders", "region", "biotoop")}}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{mapFlanders}} - UI side
}
\author{
mvarewyck
}
