% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{mapSchade}
\alias{mapSchade}
\title{Create map for Wildschade percelen}
\usage{
mapSchade(
  schadeData,
  regionLevel,
  allSpatialData,
  addGlobe = FALSE,
  legend = "topright"
)
}
\arguments{
\item{schadeData}{spatialPointsDataFrame contains the points where there was
wildschade and descriptives in data.frame}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes", "utm5" )}}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{addGlobe}{boolean, whether to add world map to background; default is FALSE}

\item{legend}{character, legend placement; default is "none", no legend}
}
\value{
leaflet map
}
\description{
Create map for Wildschade percelen
}
\author{
mvarewyck
}
