% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{mapSchadeServer}
\alias{mapSchadeServer}
\title{Shiny module for creating map on schade data - server side}
\usage{
mapSchadeServer(
  id,
  schadeData,
  allSpatialData,
  timeRange,
  defaultYear,
  species
)
}
\arguments{
\item{id}{character, unique identifier for module}

\item{schadeData}{object as returned by \code{loadRawData(type = "wildschade")}}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{timeRange}{integer vector, relevant period that can be selected for the map}

\item{defaultYear}{integer, current (default) end year of the selected period}

\item{species}{character vector, selected species for the plot}
}
\value{
no return value
}
\description{
Shiny module for creating map on schade data - server side
}
\author{
mvarewyck
}
