% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpread}
\alias{mapSpread}
\title{Create leaflet map for the spread of a species}
\usage{
mapSpread(spreadShape, legend = "none", addGlobe = FALSE)
}
\arguments{
\item{spreadShape}{SpatialPolygonsDataFrame as created by \code{createSpreadData}}

\item{legend}{character, legend placement; default is "none", no legend}

\item{addGlobe}{boolean, whether to add world map to background; default is FALSE}
}
\value{
leaflet map
}
\description{
Create leaflet map for the spread of a species
}
\author{
mvarewyck
}
