% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpreadServer}
\alias{mapSpreadServer}
\title{Shiny module for creating the plot \code{\link{mapFlanders}} - server side}
\usage{
mapSpreadServer(
  id,
  regionLevel,
  locaties,
  allSpatialData,
  type = c("F06", "F17_4"),
  title = reactive(NULL)
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{regionLevel}{reactive, pre-selected value of regionLevel chosen outside module;
to draw black borders and zoom}

\item{locaties}{reactive, pre-selected value of region chosen outside module;
to draw black borders and zoom}

\item{allSpatialData}{list with SpatialPolygonsDataFrame}

\item{type}{character, defines the layout depending on which page it is shown;
should be one of \code{c("grofwild", "wildschade", "wbe", "dash")}}

\item{title}{reactive object, title with asterisk to show in the \code{actionLink}}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{mapFlanders}} - server side
}
\author{
mvarewyck
}
