% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpreadUI}
\alias{mapSpreadUI}
\title{Shiny module for creating the plot \code{\link{mapSpread}} - UI side}
\usage{
mapSpreadUI(id, title, showLayer = FALSE)
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{mapSpread}} - UI side
}
\author{
mvarewyck
}
