% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapVerkeer}
\alias{mapVerkeer}
\title{Create leaflet map for verkeer of a species}
\usage{
mapVerkeer(
  trafficData,
  layers = c("oversteek", "ecorasters"),
  addGlobe = FALSE
)
}
\arguments{
\item{trafficData}{list, with sf data.frame for each name specified in \code{layers}}

\item{layers}{character vector,}

\item{addGlobe}{boolean, whether to add world map to background; default is FALSE}
}
\value{
leaflet map
}
\description{
Create leaflet map for verkeer of a species
}
\author{
mvarewyck
}
