% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_dashboard.R
\name{namedChoices}
\alias{namedChoices}
\title{Create named choices for the dashboard indicators}
\usage{
namedChoices(choices, uiText, regionLevel)
}
\arguments{
\item{choices}{character vector, plotFunction names in \code{uiText} 
for which to create named choices, i.e. F_* codes}

\item{uiText}{data.frame as loaded by \code{read.csv(file = file.path(dataDir, "uiText.csv"))}}

\item{regionLevel}{character, selected regional level; to define whether 
info is partially known (* indicator)}
}
\value{
named character vector
}
\description{
Create named choices for the dashboard indicators
}
\author{
mvarewyck
}
