% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{nameFile}
\alias{nameFile}
\title{Name file given content information}
\usage{
nameFile(species, year, content, fileExt)
}
\arguments{
\item{species}{character, species of the file content}

\item{year}{numeric vector, year span of the file content}

\item{content}{character, more information on the file}

\item{fileExt}{character, extension of the file}
}
\value{
character, suggested file name pasting together \code{species},
\code{year}, \code{content}, \code{fileExt}
}
\description{
Name file given content information
}
\author{
mvarewyck
}
