% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{optionsModuleServer}
\alias{optionsModuleServer}
\title{User input for controlling specific plot (server-side)}
\usage{
optionsModuleServer(input, output, session, data, types = NULL,
  timeLabel = "Periode")
}
\arguments{
\item{input}{shiny input variable for specific namespace}

\item{output}{shiny output variable for specific namespace}

\item{session}{shiny session variable for specific namespace}

\item{data}{reactive data.frame, data for chosen species}

\item{types, }{defines the species types that can be selected}

\item{timeLabel}{label for the time slider, 'Periode' by default}
}
\value{
no return value; some output objects are created
}
\description{
User input for controlling specific plot (server-side)
}
