% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{optionsModuleUI}
\alias{optionsModuleUI}
\title{User input for controlling specific plot (ui-side)}
\usage{
optionsModuleUI(
  id,
  showLegend = FALSE,
  showTime = FALSE,
  showYear = FALSE,
  showType = FALSE,
  regionLevels = NULL,
  summarizeBy = NULL,
  exportData = FALSE,
  showDataSourceSchade = FALSE,
  doWellPanel = TRUE,
  filter = FALSE,
  showDataSource = FALSE,
  showDataSourceGeslacht = FALSE,
  showInterval = FALSE
)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{showLegend}{boolean, whether to show input field for the legend}

\item{showTime}{boolean, whether to show slider input field for time range}

\item{showYear}{boolean, whether to show numeric input field for year selection}

\item{showType, }{boolean, whether to select a select input field with type}

\item{regionLevels}{numeric vector, if not NULL, defines the choices for 
region levels: 1 = flanders, 2 = provinces, 3 = communes}

\item{summarizeBy}{character, choices to be shown as summary statistics
(expect count or percent)}

\item{exportData}{boolean, whether a download button for the data is shown}

\item{showDataSourceSchade}{boolean, whether to show choices of data source to be 
used for schade data}

\item{doWellPanel}{boolean, whether to display the options within a 
\code{shiny::wellPanel()}}

\item{filter}{boolean, if TRUE percentage of data used after applying filters is shown}

\item{showDataSource}{boolean, whether to show option to select leeftijd source}

\item{showDataSourceGeslacht}{boolean, whether to show option to select geslacht source}

\item{showInterval}{boolean, if TRUE gives user option to select interval}
}
\value{
ui object (tagList)
}
\description{
User input for controlling specific plot (ui-side)
}
