% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{paletteMap}
\alias{paletteMap}
\title{Necessary info for the color palette of \code{\link{mapCube}}}
\usage{
paletteMap(variable, groupNames)
}
\arguments{
\item{groupNames}{character vector, labels to be shown in the color legend}

\item{groupVariable}{character, variable for which the \code{groupNames} are defined}
}
\value{
list with colors, character vector and levels, character vector. 
Each item has same length as \code{units}
}
\description{
Necessary info for the color palette of \code{\link{mapCube}}
}
\author{
mvarewyck
}
