% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageRealisedShotAnimals.R
\name{percentageRealisedShotServer}
\alias{percentageRealisedShotServer}
\title{Shiny module for creating the plot \code{\link{countYearRealizedShot}} - server side}
\usage{
percentageRealisedShotServer(id, data, timeRange, types)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{types}{defines the species types that can be selected}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countYearRealizedShot}} - server side
}
\author{
mvarewyck
}
