% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageRealisedShotAnimals.R
\name{percentageRealisedShotUI}
\alias{percentageRealisedShotUI}
\title{Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side}
\usage{
percentageRealisedShotUI(id, showAccuracy = FALSE)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{showAccuracy}{boolean, whether to show gauge for accuracy}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side
}
\author{
mvarewyck
}
