% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageYearlyShotAnimals.R
\name{percentageYearlyShotAnimals}
\alias{percentageYearlyShotAnimals}
\title{Create interactive plot to represent yearly percentage killed}
\usage{
percentageYearlyShotAnimals(data, wildNaam = "", type = NULL,
  doodsoorzaak = "afschot", jaartallen = NULL, jaar = NULL,
  width = NULL, height = NULL, openingstijden = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{type}{animal type}

\item{doodsoorzaak}{character vector, defines the cause(s) of death that should
be considered in the plot}

\item{jaartallen}{integer vector, defines the year(s) considered for the reference period
in the plot; if NULL no selection on year(s) is made}

\item{jaar}{year of interest}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}

\item{openingstijden}{vector with start and end of season, see output of
\link{loadOpeningstijdenData}}
}
\value{
plotly object, for a given specie the observed yearly percentage
killed animals for the year \code{jaar}. 
The mean in the entire year is represented by a dotted line.
The range of the yearly percentage killed for the \code{jaartallen} period is represented
by a ribbon, and its median by a full line.
}
\description{
Adapted version from Figure p. 13 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
