% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentCollected}
\alias{percentCollected}
\title{Generate text with percentage collected as annotation for plots}
\usage{
percentCollected(nAvailable, nTotal, text)
}
\arguments{
\item{nAvailable}{integer, number of subjects for which data is available (non-missing)}

\item{nTotal}{integer, total number of subjects}

\item{text}{character, explains how the filtering for \code{nAvailable} is done}
}
\value{
character, annotation text for the plot
}
\description{
Generate text with percentage collected as annotation for plots
}
\author{
mvarewyck
}
