% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioindicator.R
\name{plotBioindicator}
\alias{plotBioindicator}
\title{Create interactive plot for bioindicator versus year}
\usage{
plotBioindicator(data, wildNaam = "", jaartallen = NULL, regio = "",
  bioindicator = c("onderkaaklengte", "ontweid_gewicht", "aantal_embryos"),
  width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte',
'ontweid_gewicht', or 'aantal_embryos'}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object, for the specified specie and years
}
\description{
Create interactive plot for bioindicator versus year
}
\author{
Laure Cougnaud
}
