% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioindicator.R
\name{plotBioindicatorServer}
\alias{plotBioindicatorServer}
\title{Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side}
\usage{
plotBioindicatorServer(
  id,
  data,
  timeRange,
  types,
  typesDefault,
  bioindicator = c("onderkaaklengte", "ontweid_gewicht")
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{types}{defines the species types that can be selected}

\item{typesDefault}{defines the default values for \code{types},
same as \code{types} by defualt}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte' or 'ontweid_gewicht'}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side
}
\author{
mvarewyck
}
