% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plotModuleServer}
\alias{plotModuleServer}
\title{Interactive plot (server-side)}
\usage{
plotModuleServer(input, output, session, plotFunction, data,
  openingstijdenData = NULL, wildNaam)
}
\arguments{
\item{input}{shiny input variable for specific namespace}

\item{output}{shiny output variable for specific namespace}

\item{session}{shiny session variable for specific namespace}

\item{plotFunction}{character, defines the plot function to be called}

\item{data}{reactive data.frame, data for chosen species}

\item{openingstijdenData}{data with openingstijden, optional}

\item{wildNaam}{character, defines the species name for y-label in plot}
}
\value{
no return value; plot output object is created
}
\description{
Interactive plot (server-side)
}
\author{
mvarewyck
}
