% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plotModuleUI}
\alias{plotModuleUI}
\title{Interactive plot (ui-side)}
\usage{
plotModuleUI(id, height = "600px", filter = FALSE)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{height}{character, plot height, default is "600px"}

\item{filter}{boolean, whether to display filters UI}
}
\value{
ui object
}
\description{
Interactive plot (ui-side)
}
\author{
mvarewyck
}
