% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{readShapeData}
\alias{readShapeData}
\title{Read all shape data from geojson files}
\usage{
readShapeData(
  jsonDir,
  dataDir = system.file("extdata", package = "reportingGrofwild"),
  tolerance = 1e-04
)
}
\arguments{
\item{dataDir}{character, path to data files}

\item{tolerance}{numeric, defines the tolerance in the Douglas-Peuker algorithm;
larger values will impose stronger simplification; default value is 0.001}
}
\value{
save to dataDir object spatialData, i.e. a list with for each 
spatial level a SpatialPolygonsDataFrame object, 
with polygons and data as provided in the dataDir; spatial levels are 
flanders, provinces, communes, faunabeheerzones, fbz_gemeentes, utm5 
and provincesVoeren (Voeren as separate province)
}
\description{
Read all shape data from geojson files
}
