% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replicateColors}
\alias{replicateColors}
\title{Replicate inbo colors if more than 9 needed}
\usage{
replicateColors(nColors)
}
\arguments{
\item{nColors}{integer, number of colors needed}
}
\value{
list with
colors = character vector with (repeated) inbo colors
warning = character, not NULL if colors are repeated
}
\description{
Replicate inbo colors if more than 9 needed
}
\author{
mvarewyck
}
