% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{setMonthsInDutch}
\alias{setMonthsInDutch}
\title{Replace english months detected in vector elements with Dutch month names}
\usage{
setMonthsInDutch(x)
}
\arguments{
\item{x}{character or factor, what to transform}
}
\value{
character or factor vector, depending on what was supplied
}
\description{
Will not return NA, but rather the original content if
no match could be found. In case a factor is supplied, 
a factor will be returned keeping the original order of the levels, but 
with adjusted months.
}
\author{
eadriaensen
}
