% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize first letter}
\usage{
simpleCap(names, keepNames = TRUE)
}
\arguments{
\item{names}{character vector, names to be capitalized (e.g. countries)}

\item{keepNames}{boolean, whether to keep the names for the returned vector;
default is TRUE}
}
\value{
character vector, capitalized version of \code{names}
}
\description{
Capitalize first letter
}
\author{
mvarewyck
}
