% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{sourcesSchade}
\alias{sourcesSchade}
\title{List with all possible choices for indieningType per schadeChoice}
\usage{
sourcesSchade()
}
\value{
list
}
\description{
List with all possible choices for indieningType per schadeChoice
}
\author{
mvarewyck
}
