% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableBackground.R
\name{tableBackground}
\alias{tableBackground}
\title{Function to generate table for wegdensiteit (F03_1)}
\usage{
tableBackground(data, regionLevel, locaties)
}
\arguments{
\item{data}{data.frame with row for each location of interest;
columns are measured background variables}

\item{regionLevel}{reactive, pre-selected value of regionLevel chosen outside module;
to draw black borders and zoom}

\item{locaties}{reactive, pre-selected value of region chosen outside module;
to draw black borders and zoom}
}
\value{
data.frame
}
\description{
Function to generate table for wegdensiteit (F03_1)
}
\author{
wverlinden
}
