% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableGewas.R
\name{tableGewas}
\alias{tableGewas}
\title{Create summary table for gewas by region and year.}
\usage{
tableGewas(
  data,
  jaartallen = NULL,
  variable,
  sourceIndicator = NULL,
  type = c("provinces", "flanders", "faunabeheerzones")
)
}
\arguments{
\item{data}{data.frame with raw data for summary table}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{variable}{character, defines the dependent variable in the table besides \code{type}}

\item{sourceIndicator}{character, source used to filter \code{data} ('indieningType' column)
should be one of \code{c("E-loker", "HVV", "Natuurpunt")}.}

\item{type}{character, defines the region variable of interest in the table}
}
\value{
data.frame, number of opbservations per region and per \code{variable}
}
\description{
Create summary table for gewas by region and year.
}
\author{
Eva Adriaensen
}
