% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_modules.R
\name{tableModuleUI}
\alias{tableModuleUI}
\title{Interactive table (ui-side)}
\usage{
tableModuleUI(id, includeTotal = FALSE)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{includeTotal}{boolean, whether include text with total number of records in table}
}
\value{
ui object
}
\description{
Interactive table (ui-side)
}
\author{
mvarewyck
}
