% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableProvince.R
\name{tableProvince}
\alias{tableProvince}
\title{Create summary table for specific species}
\usage{
tableProvince(
  data,
  assignedData,
  jaar = NULL,
  categorie = c("leeftijd", "typeAantal", "typePercent"),
  minForTrend = 50
)
}
\arguments{
\item{data}{data.frame with raw data for summary table}

\item{assignedData}{data.frame with summary data on the number of assigned
animals that can be shot}

\item{jaar}{year of interest}

\item{categorie}{character, defines the dependent variable in the table
besides "provinces"}

\item{minForTrend}{numeric, the minimum number of records needed before
a trend is being reported}
}
\value{
data.frame, number or percentage of observations 
per province and per \code{categorie}
}
\description{
Create summary table for specific species
}
\author{
mvarewyck
}
