% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableProvince.R
\name{tableProvinceServer}
\alias{tableProvinceServer}
\title{Shiny module for creating the plot \code{\link{tableProvince}} - server side}
\usage{
tableProvinceServer(id, data, categorie, timeRange)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{categorie}{character, defines the dependent variable in the table
besides "provinces"}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{tableProvince}} - server side
}
\author{
mvarewyck
}
