% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableProvince.R
\name{tableProvinceUI}
\alias{tableProvinceUI}
\title{Shiny module for creating the plot \code{\link{tableProvince}} - UI side}
\usage{
tableProvinceUI(id, uiText)
}
\description{
Shiny module for creating the plot \code{\link{tableProvince}} - UI side
}
\author{
mvarewyck
}
