% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSpecies.R
\name{tableSpecies}
\alias{tableSpecies}
\title{Create summary table for specific species per WBE}
\usage{
tableSpecies(data, jaar = NULL, minForTrend = 10)
}
\arguments{
\item{data}{data.frame with raw data for summary table}

\item{jaar}{year of interest}

\item{minForTrend}{numeric, the minimum number of records needed before
a trend is being reported}
}
\value{
data.frame, number or percentage of observations 
per province and per \code{categorie}
}
\description{
Create summary table for specific species per WBE
}
\author{
mvarewyck
}
