% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSpecies.R
\name{tableSpeciesServer}
\alias{tableSpeciesServer}
\title{Shiny module for creating the plot \code{\link{countAgeGender}} - server side}
\usage{
tableSpeciesServer(id, data, timeRange, species)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{species}{character, species for which to show the graphs}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countAgeGender}} - server side
}
\author{
mvarewyck
}
