% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearFlanders.R
\name{trendYearFlanders}
\alias{trendYearFlanders}
\title{Create interactive plot for counts in Flanders per selected years}
\usage{
trendYearFlanders(data, timeRange, width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame, formatted data for plotting}
}
\value{
list with plot and data
}
\description{
Create interactive plot for counts in Flanders per selected years
}
\author{
mvarewyck
}
