% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{trendYearRegionUI}
\alias{trendYearRegionUI}
\title{Shiny module for creating the plot \code{\link{trendYearRegion}} - UI side}
\usage{
trendYearRegionUI(id, unitChoices)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{unitChoices, }{character vector with choices for the units}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{trendYearRegion}} - UI side
}
\author{
mvarewyck
}
