% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{vectorToTitleString}
\alias{vectorToTitleString}
\title{Transform vector ready to be in title}
\usage{
vectorToTitleString(vector)
}
\arguments{
\item{vector}{character vector with elements to be collapsed together into one string}
}
\value{
string Elements of the vector are separated by comma in the string, 
ultimate and penultimate element are separated by 'en'.
}
\description{
Separates elements of vectors with a comma; the final element of the vector is seperated with 'en'
}
\author{
Eva Adriaensen
}
