% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sections.R
\name{welcomeSection}
\alias{welcomeSection}
\title{Section for welcoming (top of the page)}
\usage{
welcomeSection(id, uiText)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Section for welcoming (top of the page)
}
\author{
mvarewyck
}
