% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageYearlyShotAnimals.R
\name{yearlyShotAnimalsServer}
\alias{yearlyShotAnimalsServer}
\title{Shiny module for creating the plot \code{\link{percentageYearlyShotAnimals}} - server side}
\usage{
yearlyShotAnimalsServer(id, data, timeRange, type, openingstijdenData)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{type}{animal type, used to filter \code{data} and \code{openingstijdenData} ('type' column)
If NULL (by default) or 'all', the data is not filtered.}

\item{openingstijdenData}{data.frame with opening season, 
as returned by the link{loadOpeningstijdenData} function}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{percentageYearlyShotAnimals}} - server side
}
\author{
mvarewyck
}
