% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageYearlyShotAnimals.R
\name{yearlyShotAnimalsUI}
\alias{yearlyShotAnimalsUI}
\title{Shiny module for creating the plot \code{\link{percentageYearlyShotAnimals}} - UI side}
\usage{
yearlyShotAnimalsUI(id, uiText)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{percentageYearlyShotAnimals}} - UI side
}
\author{
mvarewyck
}
