% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{yearToTitleString}
\alias{yearToTitleString}
\title{Transform numeric vector with years ready to be in title}
\usage{
yearToTitleString(year, brackets = TRUE)
}
\arguments{
\item{year}{numeric, vector of length 1 or 2}

\item{brackets}{logical, should the output string with years be wrapped in brackets? Defaults to TRUE}
}
\value{
string Numbers of the vector are separated by 'tot' in the string, 
brackets are put around (\code{brackets = TRUE}). If length of vector is only one, or if both elements
of the vector are the same, a string with the year between brackets is returned.
}
\description{
Transform numeric vector with years ready to be in title
}
\author{
Eva Adriaensen
}
