% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barCost.R
\name{barCost}
\alias{barCost}
\title{Function to generate stacked bar plot for kost landbouwschade (F09_2)}
\usage{
barCost(data, summarizeBy = NULL, yVar = c("schadeBedrag", "count"))
}
\arguments{
\item{data}{data.frame with schadeData}

\item{summarizeBy}{character, variable in \code{data} to summarize on}

\item{yVar}{character, column in \code{plotData} for y-axis}
}
\value{
list with plotly object and data.frame 
plot per year (xaxis) and group (color): freq x schadeBedrag (yaxis)
}
\description{
Function to generate stacked bar plot for kost landbouwschade (F09_2)
}
\author{
wverlinden
}
