% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barCost.R
\name{barCostServer}
\alias{barCostServer}
\title{Shiny module for creating the plot \code{\link{barCost}} - server side}
\usage{
barCostServer(id, yVar, data, title = reactive(NULL))
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{yVar}{character, column in \code{plotData} for y-axis}

\item{data}{data.frame for the plot function}

\item{title}{reactive, plot title to be displayed}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{barCost}} - server side
}
\author{
mvarewyck
}
