% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barCost.R
\name{barCostUI}
\alias{barCostUI}
\title{Shiny module for creating the plot \code{\link{barCost}} - UI side}
\usage{
barCostUI(id, uiText, typeMelding = NULL)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{typeMelding}{character vector, choices for filtering on `typeMelding` in data}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{barCost}} - UI side
}
\author{
mvarewyck
}
