% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barDraagkracht.R
\name{barDraagkrachtUI}
\alias{barDraagkrachtUI}
\title{Shiny module for creating the plot \code{\link{barDraagkracht}} - UI side}
\usage{
barDraagkrachtUI(
  id,
  uiText,
  sectorChoices = NULL,
  selectGroups = FALSE,
  height = "600px"
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{sectorChoices}{named character vector, choices for sector}

\item{selectGroups}{boolean, whether user should be able to select groups}

\item{height}{character, height of the plot}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{barDraagkracht}} - UI side
}
\author{
mvarewyck
}
