% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barDraagkracht.R
\name{barDraagkrachtUI}
\alias{barDraagkrachtUI}
\title{Shiny module for creating the plot \code{\link{barDraagkracht}} - UI side}
\usage{
barDraagkrachtUI(id, title, subGroups = NULL)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{title}{character, title for the plot}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{barDraagkracht}} - UI side
}
\author{
mvarewyck
}
