% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sections.R
\name{bioindicatorSection}
\alias{bioindicatorSection}
\title{Section title and text for bio-indicator}
\usage{
bioindicatorSection(id, uiText)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Section title and text for bio-indicator
}
\author{
mvarewyck
}
