% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sections.R
\name{bioindicatorSectionServer}
\alias{bioindicatorSectionServer}
\title{Section title and text for bio-indicator - server side}
\usage{
bioindicatorSectionServer(id, uiText, wildsoort)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{wildsoort}{reactive, selected wildsoort in the app}
}
\value{
no return value
}
\description{
Section title and text for bio-indicator - server side
}
\author{
mvarewyck
}
