% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxAgeWeight.R
\name{boxAgeWeight}
\alias{boxAgeWeight}
\title{Create interactive plot for distribution of weight in function of age}
\usage{
boxAgeWeight(data, type, jaartallen = NULL, regio = "", width = NULL,
  height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{type}{animal type, used to filter \code{data},
based on 'Leeftijdscategorie_onderkaak' column}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given species the distribution of weight
in function of age is plotted in box plots}
\item{'data': }{raw data used for the plot, as data.frame with:
\itemize{
\item{'gewicht': }{weight in kilograms}
\item{'leeftijd': }{age category}
\item{'maanden': }{age in months}
\item{'geslacht': }{gender}
}
}
}
}
\description{
Figure p. 17 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
