% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeGender.R
\name{countAgeGender}
\alias{countAgeGender}
\title{Create interactive plot for comparing age per gender}
\usage{
countAgeGender(data, wildNaam = "", jaartallen = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for a given species the percentage per age category
and per gender, based on meldingsformulier data in a stacked bar chart}
\item{'data': }{data displayed in the plot, as data.frame with:
\itemize{
\item{'geslacht': }{gender}
\item{'leeftijd': }{age category}
\item{'freq': }{counts of animals}
\item{'percent': }{percentage of counts of animals}
}
}
}
}
\description{
Figure p. 15 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
\author{
mvarewyck
}
