% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeGender.R
\name{countAgeGenderUI}
\alias{countAgeGenderUI}
\title{Shiny module for creating the plot \code{\link{countAgeGender}} - UI side}
\usage{
countAgeGenderUI(id, uiText)
}
\description{
Shiny module for creating the plot \code{\link{countAgeGender}} - UI side
}
\author{
mvarewyck
}
