% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAgeGroup.R
\name{countAgeGroupUI}
\alias{countAgeGroupUI}
\title{Shiny module for creating the plot \code{\link{countAgeGroup}} - UI side}
\usage{
countAgeGroupUI(id, uiText, title = NULL, groupVariable, doHide = TRUE)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{title}{character, plot title in the app, overrides the automatic title}

\item{doHide}{boolean, whether to initially hide the plot; default TRUE}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{countAgeGroup}} - UI side
}
\author{
mvarewyck
}
