% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countEmbryos.R
\name{countEmbryosUI}
\alias{countEmbryosUI}
\title{Shiny module for creating the plot \code{\link{countEmbryos}} - UI side}
\usage{
countEmbryosUI(id, regionLevels)
}
\arguments{
\item{id}{character, identifier}

\item{regionLevels}{character, choices for region}
}
\description{
Shiny module for creating the plot \code{\link{countEmbryos}} - UI side
}
\author{
mvarewyck
}
