% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoten_indicatorIntroductionYear.R
\name{countIntroductionYear}
\alias{countIntroductionYear}
\title{Plot number of new introductions per year}
\usage{
countIntroductionYear(data, region = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{region}{NULL currently not implemented}
}
\value{
list with plot and data
data that was used for the plot (i.e. without missing values for the )
}
\description{
Plot number of new introductions per year
}
