% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearAge.R
\name{countYearAgeServer}
\alias{countYearAgeServer}
\title{Shiny module for creating the plot \code{\link{countYearAge}} - server side}
\usage{
countYearAgeServer(id, data, timeRange, title = reactive(NULL))
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{title}{reactive character, title with asterisk to show in the \code{actionLink}}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{countYearAge}} - server side
}
\author{
mvarewyck
}
