% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearAge.R
\name{countYearAgeUI}
\alias{countYearAgeUI}
\title{Shiny module for creating the plot \code{\link{countYearAge}} - UI side}
\usage{
countYearAgeUI(id, uiText)
}
\description{
Shiny module for creating the plot \code{\link{countYearAge}} - UI side
}
\author{
mvarewyck
}
