% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearProvince.R
\name{countYearProvince}
\alias{countYearProvince}
\title{Create interactive plot for counts per province and year}
\usage{
countYearProvince(data, wildNaam = "", jaartallen = NULL,
  doodsoorzaak = c("afschot", "valwild"), width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{doodsoorzaak}{character vector, defines the cause(s) of death that should
be considered in the plot}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
plotly object, for a given species the observed number 
per year and per province is plotted in a stacked bar chart
}
\description{
Figure p. 4 from https://pureportal.inbo.be/portal/files/11785261/Huysentruyt_etal_2015_GrofwildjachtVlaanderen.pdf
}
