% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearProvince.R
\name{countYearProvinceUI}
\alias{countYearProvinceUI}
\title{Shiny module for creating the plot \code{\link{countYearProvince}} - UI side}
\usage{
countYearProvinceUI(
  id,
  uiText,
  plotFunction = "countYearProvinceUI",
  showType = FALSE,
  showDataSource = NULL,
  doHide = TRUE
)
}
\arguments{
\item{id}{unique identifier}

\item{plotFunction}{character, for matching uiText}

\item{doHide}{boolean, whether to initially hide the plot; default TRUE}
}
\description{
Shiny module for creating the plot \code{\link{countYearProvince}} - UI side
}
\author{
mvarewyck
}
