% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearProvince.R
\name{countYearProvinceUI}
\alias{countYearProvinceUI}
\title{Shiny module for creating the plot \code{\link{countYearProvince}} - UI side}
\usage{
countYearProvinceUI(
  id,
  uiText,
  plotFunction = "countYearProvinceUI",
  showType = FALSE,
  showDataSource = NULL,
  doHide = TRUE
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{plotFunction}{character, for matching uiText}

\item{showType}{boolean, whether to select a select input field with type}

\item{showDataSource}{character vector, for which variables to show choices 
of data source levels. 
Should be one or more of \code{c("schade", "leeftijd", "geslacht", "onderkaak", "embryos")}}

\item{doHide}{boolean, whether to initially hide the plot; default TRUE}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{countYearProvince}} - UI side
}
\author{
mvarewyck
}
