% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearShotAnimals.R
\name{countYearShotUI}
\alias{countYearShotUI}
\title{Shiny module for creating the plot \code{\link{countYearShotAnimals}} - UI side}
\usage{
countYearShotUI(id, groupVariable, regionLevels = NULL, uiText)
}
\arguments{
\item{groupVariable}{character, variable name in \code{data} for 
which colored bars should be plot}

\item{regionLevels}{numeric vector, region level choices}
}
\description{
Shiny module for creating the plot \code{\link{countYearShotAnimals}} - UI side
}
\author{
mvarewyck
}
