% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countYearShotAnimals.R
\name{countYearShotUI}
\alias{countYearShotUI}
\title{Shiny module for creating the plot \code{\link{countYearShotAnimals}} - UI side}
\usage{
countYearShotUI(id, regionLevels = NULL)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{regionLevels}{numeric vector, region level choices}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{countYearShotAnimals}} - UI side
}
\author{
mvarewyck
}
