% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{createSpaceData}
\alias{createSpaceData}
\title{Create summary data of geographical data for selected year, species and region level}
\usage{
createSpaceData(data, allSpatialData, year, species, regionLevel,
  unit = c("absolute", "relative", "absoluteCases"))
}
\arguments{
\item{data}{data.frame, geographical data}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{year}{integer, year of interest}

\item{species}{character, species of interest}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes")}}

\item{unit}{character, whether absolute or relative frequencies (aantal/100ha) 
should be reported}
}
\value{
data.frame
}
\description{
Create summary data of geographical data for selected year, species and region level
}
\author{
mvarewyck
}
