% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{createSpaceData}
\alias{createSpaceData}
\title{Create summary data of geographical data for selected year, species and region level}
\usage{
createSpaceData(data, spatialData, year, species, regionLevel)
}
\arguments{
\item{data}{data.frame, geographical data}

\item{spatialData}{sp, spatial data for the selected \code{regionLevel}}

\item{year}{integer, year of interest}

\item{species}{character, species of interest}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes")}}
}
\value{
data.frame
}
\description{
Create summary data of geographical data for selected year, species and region level
}
\author{
mvarewyck
}
