% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{createTrendData}
\alias{createTrendData}
\title{Create data for plotting trend over years in \code{\link{trendYearFlanders}}}
\usage{
createTrendData(
  data,
  allSpatialData,
  biotoopData = NULL,
  timeRange,
  species,
  regionLevel,
  unit = c("absolute", "relative", "relativeDekking"),
  sourceIndicator = NULL,
  dataDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{data}{data.frame, geographical data}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{biotoopData}{data.frame, background data for the WBE, as read from \code{loadHabitats}}

\item{timeRange}{numeric vector, time range selected for plot}

\item{species}{character, species of interest}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes", "utm5" )}}

\item{unit}{character, whether absolute frequencies, relative frequencies (aantal/100ha),
absolute cases, or relative bos freq (aantal/100ha bos & natuur); if 'region'
the legend shows different types of regions for WBE}

\item{sourceIndicator}{character, source used to filter \code{data} ('indieningType' column)
should be subset of \code{c("E-loket", "HVV", "Natuurpunt")}.}

\item{dataDir}{character, path to data files}
}
\value{
data.frame, summary of number of animals per species, region and year.
Ready for plotting with \code{\link{trendYearFlanders}}
}
\description{
Create data for plotting trend over years in \code{\link{trendYearFlanders}}
}
\author{
mvarewyck
}
