% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoten_indicatorIntroductionCumulative.R
\name{cumulativeIntroductionYear}
\alias{cumulativeIntroductionYear}
\title{Cumulative plot of yearly number of new introductions}
\usage{
cumulativeIntroductionYear(data, region = NULL)
}
\arguments{
\item{data}{}
}
\value{
plotly plot with attirubte "proportionInfo" indicating the proportion of 
data that was used for the plot (i.e. without missing values for the )
}
\description{
Cumulative plot of yearly number of new introductions
}
