% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_dashboard.R
\name{dashboardChoices}
\alias{dashboardChoices}
\title{List choices for indicatoren - UI side}
\usage{
dashboardChoices(id, choices, selected = NULL, uiText)
}
\arguments{
\item{id}{character, unique identifier for module}

\item{choices}{character vector, plotFunction names in \code{uiText} 
for which to create named choices, i.e. F_* codes}

\item{selected}{character vector, subset of \code{choices} pre-selected choices}

\item{uiText}{data.frame as loaded by \code{read.csv(file = file.path(dataDir, "uiText.csv"))}}
}
\value{
named character vector
}
\description{
List choices for indicatoren - UI side
}
\author{
mvarewyck
}
