% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_dashboard.R
\name{dashboardChoicesServer}
\alias{dashboardChoicesServer}
\title{List choices for indicatoren - server side}
\usage{
dashboardChoicesServer(id, choices, uiText, regionLevel)
}
\arguments{
\item{id}{character, unique identifier for module}

\item{choices}{character vector, indicator choices to list, i.e. F_* codes}

\item{uiText}{data.frame}

\item{regionLevel}{reactive object, selected region level}
}
\value{
selected choices by user
}
\description{
List choices for indicatoren - server side
}
\author{
mvarewyck
}
