% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadHabitats}
\alias{loadHabitats}
\title{Load Habitats (Background) data}
\usage{
loadHabitats(
  bucket = config::get("bucket", file = system.file("config.yml", package =
    "reportingGrofwild")),
  spatialData,
  regionLevels = NULL
)
}
\arguments{
\item{bucket}{character, name of the S3 bucket as specified in the config.yml file;
default value is "inbo-wbe-uat-data"}

\item{spatialData}{list with each element a SpatialPolygonsDataFrame as created 
by \code{\link{createShapeData}}}

\item{regionLevels}{character vector, for which regions load the habitat data;
if NULL loaded for all levels; default value is NULL}
}
\value{
named list with data.frame for each region level
}
\description{
Load Habitats (Background) data
}
\author{
mvarewyck
}
