% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadOpeningstijdenData}
\alias{loadOpeningstijdenData}
\title{read openingstijden data}
\usage{
loadOpeningstijdenData()
}
\value{
data.frame with openingstijden

data.frame with columns:
\itemize{
\item{'Soort': }{specie}
\item{'Type': }{specie type}
\item{'Jaar': }{year}
\item{'Startdatum': }{start datum, in the format '\%d/\%m/\%Y'}
\item{'Stopdatum': }{end datum, in the format '\%d/\%m/\%Y'}
}
}
\description{
read openingstijden data
}
