% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadOpeningstijdenData}
\alias{loadOpeningstijdenData}
\title{read openingstijden data}
\usage{
loadOpeningstijdenData(bucket = config::get("bucket"))
}
\arguments{
\item{bucket}{character, name of the S3 bucket as specified in the config.yml file;
default value is "inbo-wbe-uat-data"}
}
\value{
data.frame with columns:
\itemize{
\item{'Soort': }{specie}
\item{'Type': }{specie type}
\item{'Jaar': }{year}
\item{'Startdatum': }{start datum, in the format '\%d/\%m/\%Y'}
\item{'Stopdatum': }{end datum, in the format '\%d/\%m/\%Y'}
}
and attribute 'Date', the date that this data file was created
}
\description{
read openingstijden data
}
