% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadShapeData}
\alias{loadShapeData}
\title{Load all shape data from zipped files}
\usage{
loadShapeData()
}
\value{
list with for each spatial level a SpatialPolygonsDataFrame object, 
with polygons and data as provided in the zipFile; spatial levels are 
flanders, provinces, communes and provincesVoeren (Voeren as separate province)
}
\description{
Load all shape data from zipped files
}
