% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadToekenningen}
\alias{loadToekenningen}
\title{Read toekenningen (Ree) data}
\usage{
loadToekenningen(
  dataDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{dataDir}{character vector, defines the path to the data files}
}
\value{
data.frame with columns:
\itemize{
\item{'labeltype': }{character, type of Ree, one of \code{c("geit", "bok", "kits")}}
\item{'WBE_Naam': }{character, WBE name}
\item{'labeljaar': }{integer, year}
\item{'provincie_toek': }{character, province}
\item{'toegekend': }{integer, no. of assigned animals}
\item{'verwezenlijkt': }{integer, no. of shot animals} 
\item{'percentage_verwezenlijkt': }{numeric, percentage shot animals}
\item{'KboNummer_Toek': }{character, WBE KBO number}
}
and attribute 'Date', the date that this data file was created
}
\description{
Read toekenningen (Ree) data
}
