% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadToekenningen}
\alias{loadToekenningen}
\title{Read toekenningen (Ree) data}
\usage{
loadToekenningen(
  dataDir = system.file("extdata", package = "reportingGrofwild")
)
}
\arguments{
\item{dataDir}{character vector, defines the path to the data files}
}
\value{
data.frame with columns:
\itemize{
\item{'Provincie': }{character, province}
\item{'Jaar': }{integer, year}
\item{'Labeltype': }{character, type of Ree, one of \code{c("Geiten", "Bokken", "Kitsen")}}
\item{'Aantal': }{integer, frequencie per categorie}
}
and attribute 'Date', the date that this data file was created
}
\description{
Read toekenningen (Ree) data
}
