% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{loadToekenningen}
\alias{loadToekenningen}
\title{Read toekenningen (Ree) data}
\usage{
loadToekenningen(
  bucket = config::get("bucket", file = system.file("config.yml", package =
    "reportingGrofwild"))
)
}
\arguments{
\item{bucket}{character, name of the S3 bucket as specified in the config.yml file;
default value is "inbo-wbe-uat-data"}
}
\value{
data.frame with columns:
\itemize{
\item{'labeltype': }{character, type of Ree, one of \code{c("geit", "bok", "kits")}}
\item{'WBE_Naam': }{character, WBE name}
\item{'labeljaar': }{integer, year}
\item{'provincie_toek': }{character, province}
\item{'toegekend': }{integer, no. of assigned animals}
\item{'verwezenlijkt': }{integer, no. of shot animals} 
\item{'percentage_verwezenlijkt': }{numeric, percentage shot animals}
\item{'KboNummer_Toek': }{character, WBE KBO number}
}
and attribute 'Date', the date that this data file was created
}
\description{
Read toekenningen (Ree) data
}
