% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{mapFlanders}
\alias{mapFlanders}
\title{Create map for Flanders - color by incidence}
\usage{
mapFlanders(
  regionLevel = c("flanders", "provinces", "communes", "faunabeheerzones",
    "fbz_gemeentes"),
  species,
  allSpatialData,
  summaryData,
  colorScheme,
  legend = "none",
  addGlobe = FALSE
)
}
\arguments{
\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes")}}

\item{species}{character, species of interest}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{summaryData}{data.frame, as returned by \code{\link{createSpaceData}}}

\item{colorScheme}{character vector, specifies the color palette for the different groups
in the summary data}

\item{legend}{character, legend placement; default is "none", no legend}

\item{addGlobe}{boolean, whether to add world map to background; default is FALSE}
}
\value{
leaflet map
}
\description{
Create map for Flanders - color by incidence
}
\author{
mvarewyck
}
