% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{mapFlandersServer}
\alias{mapFlandersServer}
\title{Shiny module for creating the plot \code{\link{mapFlanders}} - server side}
\usage{
mapFlandersServer(
  id,
  defaultYear,
  species,
  currentWbe = NULL,
  hideGlobeDefault = TRUE,
  type = c("grofwild", "wildschade", "wbe", "empty"),
  geoData,
  biotoopData = NULL,
  allSpatialData,
  regionLevel = reactive(NULL),
  locaties = reactive(NULL),
  countVariable = NULL,
  uiText
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{defaultYear}{numeric, default year}

\item{species}{character, species for which to show the graphs}

\item{currentWbe}{numeric, KBO number; default value is NULL}

\item{hideGlobeDefault}{boolean, whether the globe is shown by default 
when the map is first created; default value is TRUE}

\item{type}{character, defines the layout depending on which page it is shown;
should be one of \code{c("grofwild", "wildschade", "wbe", "dash")}}

\item{geoData}{SpatialPolygonsDataFrame with geographical data}

\item{biotoopData}{data.frame, with background biotoop data for selected region level;
default value is NULL}

\item{allSpatialData}{list with SpatialPolygonsDataFrame}

\item{regionLevel}{reactive, pre-selected value of regionLevel chosen outside module;
to draw black borders and zoom}

\item{locaties}{reactive, pre-selected value of region chosen outside module;
to draw black borders and zoom}

\item{countVariable}{character, column name in \code{data} that contains counts;
if NULL then each row in the data contains 1 count}

\item{uiText}{data.frame}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{mapFlanders}} - server side
}
\author{
mvarewyck
}
