% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\name{mapFlandersUI}
\alias{mapFlandersUI}
\title{Shiny module for creating the plot \code{\link{mapFlanders}} - UI side}
\usage{
mapFlandersUI(
  id,
  showRegion = TRUE,
  showCombine = TRUE,
  plotDetails = c("flanders", "region")
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{showRegion}{boolean, whether to show choices for regionLevel and selected region(s)}

\item{showCombine}{boolean, whether to show option to combine selected regions}

\item{plotDetails}{character vector, detail plots to be shown below the map;
should be subset of \code{c("flanders", "region", "biotoop")}}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{mapFlanders}} - UI side
}
\author{
mvarewyck
}
