% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{mapSchadeServer}
\alias{mapSchadeServer}
\title{Shiny module for creating map on schade data - server side}
\usage{
mapSchadeServer(
  id,
  schadeData,
  allSpatialData,
  timeRange,
  defaultYear,
  species,
  borderRegion = NULL,
  type = c("schade", "afschot")
)
}
\arguments{
\item{id}{character, unique identifier for module}

\item{schadeData}{reacive object as returned by \code{loadRawData(type = "wildschade")}}

\item{allSpatialData}{reactive object spatialPolygonsDataFrame with
spatial data for selected region (and year for WBE)}

\item{timeRange}{integer vector, relevant period that can be selected for the map}

\item{defaultYear}{integer, current (default) end year of the selected period}

\item{species}{character vector, selected species for the plot}

\item{borderRegion}{character, for which \code{regionLevel} to show black border;
see also \code{\link{mapSchade}}; default is NULL}

\item{type}{character, type of plot this module is used for. Historically
only \code{"schade"}. Later extended to also cover \code{"afschot"}, i.e. 
"gerapporteerde afschot locaties", see also \code{mapAfschotUI}}
}
\value{
no return value
}
\description{
Shiny module for creating map on schade data - server side
}
\author{
mvarewyck
}
