% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{mapSchadeUI}
\alias{mapSchadeUI}
\title{Shiny module for creating map on schade data - UI side}
\usage{
mapSchadeUI(
  id,
  filterCode = FALSE,
  filterSubcode = FALSE,
  filterSource = TRUE,
  filterAccuracy = FALSE,
  variableChoices = c(Seizoen = "season", Jaar = "afschotjaar", `Type schade` =
    "schadeCode"),
  uiText,
  plotDetails = NULL
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{filterCode}{boolean whether to include the option to filter on schade code;
default value is FALSE}

\item{filterSubcode}{boolean, whether to include the option to filter on schade subcode;
default value is FALSE}

\item{filterSource}{boolean, whether to show filter option for source}

\item{filterAccuracy}{boolean, whether to show filter option for accuracy}

\item{variableChoices}{named character vector, choices for coloring variable}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{plotDetails}{character vector, detail plots to be shown below the map;
should be subset of \code{c("flanders", "region", "biotoop")}}
}
\value{
HTML object
}
\description{
Shiny module for creating map on schade data - UI side
}
\author{
mvarewyck
}
