% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSchade.R
\name{mapSchadeUI}
\alias{mapSchadeUI}
\title{Shiny module for creating map on schade data - UI side}
\usage{
mapSchadeUI(
  id,
  filterCode = FALSE,
  filterSubcode = FALSE,
  uiText,
  plotDetails = NULL
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{filterCode}{boolean whether to include the option to filter on schade code;
default value is FALSE}

\item{filterSubcode}{boolean, whether to include the option to filter on schade subcode;
default value is FALSE}

\item{plotDetails}{character vector, detail plots to be shown below the map;
should be subset of \code{c("flanders", "region", "biotoop")}}
}
\description{
Shiny module for creating map on schade data - UI side
}
\author{
mvarewyck
}
