% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpread}
\alias{mapSpread}
\title{Create leaflet map for the spread of a species}
\usage{
mapSpread(
  spatialDir = system.file("extdata", package = "reporting-grofwild"),
  spatialLevel = c("pixels", "municipalities"),
  unit = c("model_EP", "model_OH", "risk_EP", "risk_OH"),
  startYear = 2019,
  legend = "none",
  addGlobe = FALSE
)
}
\arguments{
\item{spatialDir}{path to directory with spatial files}

\item{spatialLevel}{character, for which spatial level to create the map;
should be one of \code{c("pixels", "municipalities")}}

\item{unit}{character, characteristic that defines the polygons and color coding
should be one of \code{c("model_EP", "model_OH", "risk_EP", "risk_OH")}}

\item{startYear}{integer, starting year; only relevant for pixels & model unit}

\item{legend}{character, legend placement; default is "none", no legend}

\item{addGlobe}{boolean, whether to add world map to background; default is FALSE}
}
\value{
leaflet map
}
\description{
Create leaflet map for the spread of a species
}
\author{
mvarewyck
}
