% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpreadServer}
\alias{mapSpreadServer}
\title{Shiny module for creating the plot \code{\link{mapFlanders}} - server side}
\usage{
mapSpreadServer(
  id,
  regionLevel,
  locaties,
  allSpatialData,
  type = c("F06", "F17_4"),
  title = reactive(NULL)
)
}
\arguments{
\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes", "utm5" )}}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{title}{reactive object, title with asterisk to show in the \code{actionLink}}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{mapFlanders}} - server side
}
\author{
mvarewyck
}
