% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpreadUI}
\alias{mapSpreadUI}
\title{Shiny module for creating the plot \code{\link{mapSpread}} - UI side}
\usage{
mapSpreadUI(id, uiText, showLayer = FALSE)
}
\arguments{
\item{showLayer}{boolean, whether to display option to choose layer}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{mapSpread}} - UI side
}
\author{
mvarewyck
}
