% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{mapSpreadUI}
\alias{mapSpreadUI}
\title{Shiny module for creating the plot \code{\link{mapSpread}} - UI side}
\usage{
mapSpreadUI(id, uiText, showLayer = FALSE)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{showLayer}{boolean, whether to show the option to select spatial layers}
}
\value{
HTML object

UI object
}
\description{
Shiny module for creating the plot \code{\link{mapSpread}} - UI side
}
\author{
mvarewyck
}
