% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{namedChoices}
\alias{namedChoices}
\title{Create named choices for the dashboard indicators}
\usage{
namedChoices(choices, uiText)
}
\arguments{
\item{choices}{character vector, plotFunction names in \code{uiText} 
for which to create named choices}

\item{uiText}{data.frame as loaded by \code{read.csv(file = file.path(dataDir, "uiText.csv"))}}
}
\value{
named character vector
}
\description{
Create named choices for the dashboard indicators
}
\author{
mvarewyck
}
