% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{optionsModuleServer}
\alias{optionsModuleServer}
\title{User input for controlling specific plot (server-side)}
\usage{
optionsModuleServer(
  input,
  output,
  session,
  data,
  types = NULL,
  labelTypes = "Type",
  typesDefault = types,
  timeRange = NULL,
  timeLabel = "Periode",
  multipleTypes = FALSE,
  definedYear = defaultYear,
  sources = NULL,
  sourceLabel = "Data bron",
  sourceVariable_schade = NULL,
  sourceVariable = NULL,
  sourceVariable_geslacht = NULL,
  intervals = NULL
)
}
\arguments{
\item{input}{shiny input variable for specific namespace}

\item{output}{shiny output variable for specific namespace}

\item{session}{shiny session variable for specific namespace}

\item{data}{reactive data.frame, data for chosen species}

\item{types, }{defines the species types that can be selected}

\item{labelTypes}{character, the displayed label for selecting options field}

\item{typesDefault, }{defines the default values for \code{types},
same as \code{types} by defualt}

\item{timeRange}{numeric vector of length 2 with time range (in year)}

\item{timeLabel}{character, label for the time slider, 'Periode' by default}

\item{multipleTypes}{boolean, whether multiple types can be selected or not}

\item{definedYear}{numeric, single numeric value specifying the year value 
(or max year value within a range) that is selected upon opening, default is
\code{defaultYear} which is globally defined as \code{currentYear - 1}}

\item{sources}{defines the data sources that can be selected}

\item{sourceLabel}{character, the displayed label for the selecting source field for leeftijd, 
'Data bron' by default}

\item{sourceVariable_schade}{character, selected levels to filter for \code{indieningType} for schadeData}

\item{sourceVariable}{character, variable to filter for \code{data}}

\item{sourceVariable_geslacht}{character, selected levels to filter for \code{geslacht_comp_bron}}

\item{intervals}{defines the intervals that can be selected}
}
\value{
no return value; some output objects are created
}
\description{
User input for controlling specific plot (server-side)
}
