% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{optionsModuleServer}
\alias{optionsModuleServer}
\title{User input for controlling specific plot (server-side)}
\usage{
optionsModuleServer(input, output, session, data)
}
\arguments{
\item{input}{shiny input variable for specific namespace}

\item{output}{shiny output variable for specific namespace}

\item{session}{shiny session variable for specific namespace}

\item{data}{reactive data.frame, data for chosen species}
}
\value{
no return value; some output objects are created
}
\description{
User input for controlling specific plot (server-side)
}
