% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{optionsModuleUI}
\alias{optionsModuleUI}
\title{User input for controlling specific plot (ui-side)}
\usage{
optionsModuleUI(id, showLegend = FALSE, showExtraVariables = FALSE,
  showTime = FALSE, showRegion = FALSE, showGlobe = FALSE)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{showLegend}{boolean, whether to show input field for the legend}

\item{showExtraVariables}{boolean, whether to show input field for 
extra variables (popup)}

\item{showTime}{boolean, whether to show input field for time range}

\item{showRegion}{boolean, whether to show input fields for region selection}

\item{showGlobe}{boolean, whether to show input field for map of the globe}
}
\value{
ui object (tagList)
}
\description{
User input for controlling specific plot (ui-side)
}
