% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{optionsModuleUI}
\alias{optionsModuleUI}
\title{User input for controlling specific plot (ui-side)}
\usage{
optionsModuleUI(id, showLegend = FALSE, showTime = FALSE,
  showYear = FALSE, showType = FALSE, regionLevels = NULL,
  showSummarizeBy = FALSE, exportData = FALSE, showDataSource = FALSE)
}
\arguments{
\item{id}{character, module id, unique name per plot}

\item{showLegend}{boolean, whether to show input field for the legend}

\item{showTime}{boolean, whether to show slider input field for time range}

\item{showYear}{boolean, whether to show numeric input field for year selection}

\item{showType, }{boolean, whether to select a select input field with type}

\item{regionLevels}{numeric vector, if not NULL, defines the choices for 
region levels: 1 = flanders, 2 = provinces, 3 = communes}

\item{showSummarizeBy}{boolean, whether to show input field to choose between
aantal of percentage}

\item{exportData}{boolean, whether a download button for the data is shown}

\item{showDataSource}{boolean, whether to show choices of data source to be 
used for plotted bioindicator}
}
\value{
ui object (tagList)
}
\description{
User input for controlling specific plot (ui-side)
}
