% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpread.R
\name{paletteMap}
\alias{paletteMap}
\title{Necessary info for the color palette of \code{\link{mapSpread}}}
\usage{
paletteMap(variable, groupNames)
}
\arguments{
\item{variable}{character, data variable for which the \code{groupNames} are defined
and that will be used for coloring}

\item{groupNames}{character, factor levels that will be used as labels for colors}
}
\value{
list with colors, character vector and levels, character vector. 
Each item has same length as \code{units}
}
\description{
Necessary info for the color palette of \code{\link{mapSpread}}
}
\author{
mvarewyck
}
