% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageRealisedShotAnimals.R
\name{percentageRealisedShotUI}
\alias{percentageRealisedShotUI}
\title{Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side}
\usage{
percentageRealisedShotUI(id, showAccuracy = FALSE, uiText)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{showAccuracy}{boolean, whether to show gauge for accuracy}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side
}
\author{
mvarewyck
}
