% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioindicator.R
\name{plotBioindicator}
\alias{plotBioindicator}
\title{Create interactive plot for bioindicator versus year}
\usage{
plotBioindicator(data, wildNaam = "", type = NULL, jaartallen = NULL,
  regio = "", bioindicator = c("onderkaaklengte", "ontweid_gewicht",
  "aantal_embryos"), width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{wildNaam}{character, the name that should be displayed for y-label}

\item{type}{animal type, used to filter \code{data} ('ageGender' column)
If NULL (by default), for bioindicator set to:
\itemize{
\item{'aantal_embryos': }{only animals with age/gender type: "Smalree" "Geit" are considered}
\item{'ontweid_gewicht' or 'onderkaaklengte'}{only animals without age/gender type are filtered (\code{ageGender} set to '')}
}}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte',
'ontweid_gewicht', or 'aantal_embryos'}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for the specified specie and years}
\item{'data': }{
\itemize{
\item{for bioindicator set to 'aantal_embryos': }{
data displayed in the plot, as data.frame with:
\itemize{
\item{'afschotjaar': }{year at which the animal was shot}
\item{'variable': }{age/gender category}
\item{'Freq': }{counts of females}
}}
\item{for bioindicator set to 'ontweid_gewicht' or 'onderkaaklengte': }{
raw data used for the plot, as data.frame with:
\itemize{
\item{'afschotjaar': }{year at which the animal was shot}
\item{'variable': }{value of the bioindicator, a.k.a
weight for 'ontweid_gewicht' or length of the lower jaw for
'onderkaaklengte'}}
}}
}
}
}
\description{
Create interactive plot for bioindicator versus year
}
\author{
Laure Cougnaud
}
