% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioindicator.R
\name{plotBioindicator}
\alias{plotBioindicator}
\title{Create interactive plot for bioindicator versus year}
\usage{
plotBioindicator(
  data,
  type = NULL,
  jaartallen = NULL,
  regio = "",
  bioindicator = c("onderkaaklengte", "ontweid_gewicht"),
  sourceIndicator = c("inbo", "meldingsformulier", "both"),
  sourceIndicator_leeftijd = NULL,
  sourceIndicator_geslacht = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame with raw data for plotting}

\item{type}{animal type, used to filter \code{data} ('type_comp' column)}

\item{jaartallen}{integer vector, defines the year(s) that should be considered
in the plot; if NULL no selection on year(s) is made}

\item{regio}{character vector, names of the selected regions in \code{data}
to be shown in the plot title}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte' or 'ontweid_gewicht'}

\item{sourceIndicator}{character, defines the data source to be used for
bioindicator 'onderkaak_comp_bron'}

\item{sourceIndicator_leeftijd}{character, levels of \code{leeftijd_comp_bron} to select on}

\item{sourceIndicator_geslacht}{character, levels of \code{geslacht_comp_bron} to select on}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with:
\itemize{
\item{'plot': }{plotly object, for the specified specie and years}
\item{'data': }{
\itemize{
\item{for bioindicator set to 'ontweid_gewicht' or 'onderkaaklengte': }{
raw data used for the plot, as data.frame with:
\itemize{
\item{'afschotjaar': }{year at which the animal was shot}
\item{'variable': }{value of the bioindicator, a.k.a
weight for 'ontweid_gewicht' or length of the lower jaw for
'onderkaaklengte'}}
}}
}
}
}
\description{
Create interactive plot for bioindicator versus year
}
\author{
Laure Cougnaud
}
