% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioindicator.R
\name{plotBioindicatorUI}
\alias{plotBioindicatorUI}
\title{Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side}
\usage{
plotBioindicatorUI(
  id,
  bioindicator = c("onderkaaklengte", "ontweid_gewicht"),
  regionLevels,
  uiText
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte',
'ontweid_gewicht', or 'aantal_embryos'}

\item{regionLevels}{numeric vector, if not NULL, defines the choices for 
region levels: 1 = flanders, 2 = provinces, 3 = communes}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side
}
\author{
mvarewyck
}
