% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioindicator.R
\name{plotBioindicatorUI}
\alias{plotBioindicatorUI}
\title{Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side}
\usage{
plotBioindicatorUI(id, bioindicator = c("onderkaaklengte", "ontweid_gewicht"))
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte',
'ontweid_gewicht', or 'aantal_embryos'}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{plotBioindicator}} - UI side
}
\author{
mvarewyck
}
