% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_modules.R
\name{plotModuleServer}
\alias{plotModuleServer}
\title{Interactive plot or table (server-side)}
\usage{
plotModuleServer(
  input,
  output,
  session,
  plotFunction,
  data,
  openingstijdenData,
  toekenningsData = NULL,
  categorie = NULL,
  bioindicator = NULL,
  groupVariable = NULL,
  xVar = NULL,
  yVar = NULL,
  locaties = NULL,
  timeRange = NULL,
  unit = NULL,
  isSchade = NULL,
  datatable = FALSE,
  schadeChoices = NULL,
  schadeChoicesVrtg = NULL,
  schadeChoicesGewas = NULL,
  variable = NULL,
  combinatie = NULL,
  title = NULL,
  verticalGroups = NULL,
  fullNames = NULL
)
}
\arguments{
\item{input}{shiny input variable for specific namespace}

\item{output}{shiny output variable for specific namespace}

\item{session}{shiny session variable for specific namespace}

\item{plotFunction}{character, defines the plot function to be called}

\item{data}{reactive data.frame, data for chosen species}

\item{openingstijdenData}{data with openingstijden, optional}

\item{toekenningsData}{data with toekenningen, optional}

\item{categorie}{character, defines which type of table should be made}

\item{bioindicator}{string with column of \code{data}
with bioindicator, either 'onderkaaklengte' or 'ontweid_gewicht'}

\item{groupVariable}{character, variable name in \code{data} for 
which colored bars should be plot}

\item{xVar}{character, column in \code{plotData} for x-axis;
default value is 'percentage'}

\item{yVar}{character, column in \code{plotData} for y-axis}

\item{locaties}{character, defines on which locations to filter on;
defined externally for large map}

\item{timeRange}{numeric vector, defines on which year range to filter on;
defined externally for large map}

\item{unit}{character, defines whether absolute or relative frequencies are reported;
defined externally for large map}

\item{isSchade}{boolean, indicates whether module is used for schadeData; default is NULL}

\item{datatable}{boolean, indicates whether module should be used to output a datatable object for table; default is FALSE}

\item{schadeChoices}{character, chosen schade types (basisCode) to filter on, optional}

\item{schadeChoicesVrtg}{character, chosen schade types related to "VRTG" to filter on, optional}

\item{schadeChoicesGewas}{character, chosen schade types related to "GEWAS" to filter on, optional}

\item{variable}{character, defines which variable is of interest for the table}

\item{combinatie}{logical, summarised view of selected regions}

\item{title}{reactive, plot title to be displayed}

\item{verticalGroups}{}

\item{fullNames}{named character vector, values for the \code{variable} to be 
displayed instead of original data values}
}
\value{
no return value; plot output object is created
}
\description{
Interactive plot or table (server-side)
}
\author{
mvarewyck
}
