% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{runWildApp}
\alias{runWildApp}
\title{Run the reporting-grofwild application}
\usage{
runWildApp(installDependencies = FALSE, copyFiles = FALSE, public = TRUE, ...)
}
\arguments{
\item{installDependencies}{boolean, whether to first install packages listed
in the Suggests field of DESCRIPTION; default value is FALSE}

\item{copyFiles}{boolean, whether to copy files for the app from the inst/ui
to a temporary directory which is writeable}

\item{public}{boolean, whether to start the public or private version of the app}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Run the reporting-grofwild application
}
