% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{runWildApp}
\alias{runWildApp}
\title{Run the reporting-grofwild application}
\usage{
runWildApp(installDependencies = FALSE, public = TRUE, kbo = "", ...)
}
\arguments{
\item{installDependencies}{boolean, whether to first install packages listed
in the Suggests field of DESCRIPTION; default value is FALSE}

\item{public}{boolean, whether to start the public or private version of the app}

\item{kbo}{numeric, specific KBO number for which to show the WBE (private) app;
only relevant if \code{public} is FALSE}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Run the reporting-grofwild application
}
