% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize first letter}
\usage{
simpleCap(names)
}
\arguments{
\item{names}{character vector, names to be capitalized (e.g. countries)}
}
\value{
character vector, capitalized version of \code{names}
}
\description{
Capitalize first letter
}
\author{
mvarewyck
}
