% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableBackground.R
\name{tableBackground}
\alias{tableBackground}
\title{Function to generate table for wegdensiteit (F03_1)}
\usage{
tableBackground(biotoopData, locaties)
}
\arguments{
\item{biotoopData}{data.frame}

\item{locations}{character vector, selected regions}

\item{character}{vector containing names of provinces to take into account}
}
\value{
data.frame
}
\description{
Function to generate table for wegdensiteit (F03_1)
}
\author{
wverlinden
}
