% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{tableModuleUI}
\alias{tableModuleUI}
\title{Interactive table (ui-side)}
\usage{
tableModuleUI(id)
}
\arguments{
\item{id}{character, module id, unique name per plot}
}
\value{
ui object
}
\description{
Interactive table (ui-side)
}
\author{
mvarewyck
}
