% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSchadeCode.R
\name{tableSchadeCode}
\alias{tableSchadeCode}
\title{Create summary table for schadeCode by region.
The table created is a datatable.}
\usage{
tableSchadeCode(
  data,
  jaartallen = NULL,
  type = c("provinces", "flanders", "faunabeheerzones")
)
}
\arguments{
\item{data}{data.frame with raw data for summary table}

\item{type}{character, defines the region variable of interest in the table}
}
\value{
data.frame, number of observations per region \code{locaties} and per schadeCode
}
\description{
Create summary table for schadeCode by region.
The table created is a datatable.
}
\author{
Eva Adriaensen
}
