% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSpecies.R
\name{tableSpeciesUI}
\alias{tableSpeciesUI}
\title{Shiny module for creating the plot \code{\link{countAgeGender}} - UI side}
\usage{
tableSpeciesUI(id)
}
\arguments{
\item{id}{character, unique identifier for the module}
}
\value{
UI object
}
\description{
Shiny module for creating the plot \code{\link{countAgeGender}} - UI side
}
\author{
mvarewyck
}
