% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearFlanders.R
\name{trendYearFlanders}
\alias{trendYearFlanders}
\title{Create interactive plot for counts in Flanders per selected years}
\usage{
trendYearFlanders(
  data,
  timeRange,
  unit = c("absolute", "relative", "relativeDekking"),
  isSchade = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data.frame, formatted data for plotting as returned by \code{\link{createTrendData}}}

\item{timeRange}{numeric vector, time range selected for plot}

\item{unit}{character, whether absolute frequencies, relative frequencies (aantal/100ha),
absolute cases, or relative bos freq (aantal/100ha bos & natuur); if 'region'
the legend shows different types of regions for WBE
should be reported,}

\item{isSchade}{boolean, whether the function should generate titles for schadeData; default is FALSE}

\item{width}{plot width (optional)}

\item{height}{plot height (optional)}
}
\value{
list with plot and data
}
\description{
Create interactive plot for counts in Flanders per selected years
}
\author{
mvarewyck
}
