% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{trendYearRegionServer}
\alias{trendYearRegionServer}
\title{Shiny module for creating the plot \code{\link{trendYearRegion}} - server side}
\usage{
trendYearRegionServer(
  id,
  species,
  allSpatialData,
  biotoopData,
  geoData,
  type = "wbe",
  regionLevelName
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{geoData}{reactive data.frame, geographical data for the selected species}

\item{type}{character, type of module e.g. "wbe"}

\item{regionLevelName}{character, region name to be shown in the plot title}

\item{data}{reactive data.frame for the plot function}
}
\value{
no return value
}
\description{
NOTE: Currently only used for type = "wbe"
}
\author{
mvarewyck
}
