% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{trendYearRegionServer}
\alias{trendYearRegionServer}
\title{Shiny module for creating the plot \code{\link{trendYearRegion}} - server side}
\usage{
trendYearRegionServer(
  id,
  data,
  timeRange = reactive(NULL),
  species,
  regionLevel = reactive("WBE_buitengrenzen"),
  locaties,
  geoData,
  allSpatialData,
  biotoopData = NULL,
  title = reactive(NULL),
  type = "wbe"
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{data}{data.frame for the plot function}

\item{timeRange}{numeric vector of length 2, min and max year to subset data}

\item{species}{character, species of interest}

\item{regionLevel}{character, regional level of interest should be one of 
\code{c("flanders", "provinces", "communes", "faunabeheerzones", "fbz_gemeentes", "utm5" )}}

\item{locaties}{reactive character vector, region name to be shown in the plot title}

\item{geoData}{reactive data.frame, geographical data for the selected species}

\item{allSpatialData}{list of sp, spatial data for all spatial levels}

\item{biotoopData}{data.frame, background data for the WBE, as read from \code{loadHabitats}}

\item{title}{reactive character, title with asterisk to show in the \code{actionLink}}

\item{type}{character, type of module e.g. "wbe"}
}
\value{
no return value
}
\description{
Shiny module for creating the plot \code{\link{trendYearRegion}} - server side
}
\author{
mvarewyck
}
