% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{trendYearRegionUI}
\alias{trendYearRegionUI}
\title{Shiny module for creating the plot \code{\link{trendYearRegion}} - UI side}
\usage{
trendYearRegionUI(id, unitChoices)
}
\arguments{
\item{id}{unique identifier}

\item{unitChoices, }{character vector with choices for the units}

\item{unitChoices}{character vector, user choices for unit}
}
\description{
Shiny module for creating the plot \code{\link{trendYearRegion}} - UI side
}
\author{
mvarewyck
}
