% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendYearRegion.R
\name{trendYearRegionUI}
\alias{trendYearRegionUI}
\title{Shiny module for creating the plot \code{\link{trendYearRegion}} - UI side}
\usage{
trendYearRegionUI(
  id,
  uiText,
  plotFunction = "trendYearRegionUI",
  showCombinatie = FALSE,
  doHide = TRUE,
  unitChoices = c(Aantal = "absolute", `Aantal/100ha` = "relative",
    `Aantal/100ha bos & natuur` = "relativeDekking")
)
}
\arguments{
\item{id}{character, unique identifier for the module}

\item{uiText}{data.frame, HTML formatted text to be displayed in the UI}

\item{plotFunction}{character, for matching uiText}

\item{showCombinatie}{boolean, whether to show the option to combine lines}

\item{doHide}{boolean, whether to initially hide the plot; default TRUE}

\item{unitChoices, }{character vector with choices for the units}
}
\value{
HTML object
}
\description{
Shiny module for creating the plot \code{\link{trendYearRegion}} - UI side
}
\author{
mvarewyck
}
