% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sections.R
\name{welcomeSection}
\alias{welcomeSection}
\title{Section for welcoming (top of the page)}
\usage{
welcomeSection(id, uiText)
}
\arguments{
\item{id}{character, from which page this function is called
e.g. 'wbe'}
}
\description{
Section for welcoming (top of the page)
}
\author{
mvarewyck
}
