% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageYearlyShotAnimals.R
\name{yearlyShotAnimalsUI}
\alias{yearlyShotAnimalsUI}
\title{Shiny module for creating the plot \code{\link{percentageYearlyShotAnimals}} - UI side}
\usage{
yearlyShotAnimalsUI(id, uiText)
}
\description{
Shiny module for creating the plot \code{\link{percentageYearlyShotAnimals}} - UI side
}
\author{
mvarewyck
}
